% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_management.R
\name{add_source}
\alias{add_source}
\title{Add source to CSV file}
\usage{
add_source(source, csv_file = "data_raw/sources.csv")
}
\arguments{
\item{source}{Source object to add}

\item{csv_file}{Path to CSV file}
}
\value{
Invisible NULL
}
\description{
Add source to CSV file
}
