% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priogrid.R
\name{base_assertions}
\alias{base_assertions}
\title{base_assertions}
\usage{
base_assertions(rast)
}
\arguments{
\item{rast}{A raster layer.}
}
\value{
The function returns a character vector of messages with the results
of the assertions.
}
\description{
Applies a basic battery of assertions to the provided grid, making sure it
has the correct size, projection, resolution, in relation to PRIOGrid.
}
