% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{calculate_pgvariables}
\alias{calculate_pgvariables}
\title{Calculates all PRIO-GRID variables}
\usage{
calculate_pgvariables(gen_functions = NULL, overwrite = FALSE)
}
\arguments{
\item{gen_functions}{Character vector with priogrid functions starting with gen_. If NULL, then all are used.}

\item{overwrite}{Boolean, if false, will ignore updating variables that already exist in the result folder}
}
\description{
In PRIO-GRID, all functions that are named "gen_something()" are functions that
return final variables as a SpatRaster (or a stack of these). This function
calculates all variables and store them in "path/to/your/rawfolder/priogrid/version/{options_hash}/variable_name.rds".
}
\examples{
\dontrun{
  calculate_pgvariables()
  r <- readRDS(file.path(pgoptions$get_rawfolder(), "priogrid", packageVersion("priogrid"), "naturalearth_cover.rds"))
}
}
