% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priogrid.R
\name{check_assertions}
\alias{check_assertions}
\title{check_assertions}
\usage{
check_assertions(rast, assertions)
}
\arguments{
\item{rast}{A raster layer.}

\item{assertions}{A character vector of valid R expressions.}
}
\value{
The function returns a character vector of messages with the results
of the assertions.
}
\description{
Applies a custom battery of assertions to the provided raster layer, specified 
as a list of expressions in the config file. Each expression must evaluate to either  
true or false. The result is returned as a series of messages.
}
