% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity.R
\name{check_folder_integrity}
\alias{check_folder_integrity}
\title{check_folder_integrity}
\usage{
check_folder_integrity(folder, variable)
}
\arguments{
\item{folder}{The folder to check}

\item{variable}{The variable from which to get the hash value (to check)}
}
\description{
Checks that the contents of a folder correspond to a hash fingerprint.
}
