% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{check_pgsourcefiles}
\alias{check_pgsourcefiles}
\title{Test if MD5 checksums of local files are the same as a tested set of files}
\usage{
check_pgsourcefiles(verbose = TRUE)
}
\value{
data.frame
}
\description{
Here, we use \link{pgchecksum}, which we created when testing
PRIO-GRID, and test it against a similar method for your own
local files. This is to verify that you are using the same files
as we used to build PRIO-GRID.
}
\examples{
res <- check_pgsourcefiles()
}
