% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffunctions.R
\name{compare_crossection}
\alias{compare_crossection}
\title{Map changes}
\usage{
compare_crossection(crossection_date, cshp)
}
\description{
Determines which months a set of geometries has changed 
Used to determine when to reraster, such as in
make_gwcodes.
}
