% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_pg_indices}
\alias{create_pg_indices}
\title{Create matrix with index numbering conventions as for PRIO-GRID.}
\usage{
create_pg_indices(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link[=pg_current_config]{pg_current_config()}}.}
}
\value{
A ncol*nrow matrix with integer indices.
}
\description{
Create matrix with index numbering conventions as for PRIO-GRID.
}
\examples{
pg <- create_pg_indices()
}
