% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{cshapes_cover}
\alias{cshapes_cover}
\title{Whether or not a grid-cell intersects with the international state system (cShapes 2.0)}
\usage{
cshapes_cover(measurement_date, min_cover = 0, cshp = read_cshapes())
}
\arguments{
\item{measurement_date}{A single date as Date object}

\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\description{
Takes the cShapes data and returns
a raster-mask that is true for the grid cells that intersects with country borders
included in the international state system at the measurement_date
}
\examples{
# cshapes_cover_one_cross_section <- cshapes_cover(as.Date("2010-01-01"))
}
\references{
\insertRef{schvitzMappingInternationalSystem2022}{priogrid}
}
