% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{cshapes_gwcode}
\alias{cshapes_gwcode}
\title{The Gleditsch-Ward (cShapes 2.0 version) code in each grid-cell}
\usage{
cshapes_gwcode(measurement_date, cshp = read_cshapes())
}
\arguments{
\item{measurement_date}{A single date as Date object}

\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\value{
SpatRast
}
\description{
Uses max area for cells with multiple countries. If actual tie, the algorithm will use the lowest index.
}
\examples{
# gwcode_one_cross_section <- cshapes_gwcode(as.Date("2010-01-01"))
}
