% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priogrid.R
\name{dovar}
\alias{dovar}
\title{dovar}
\usage{
dovar(path, output_folder, variable_config)
}
\arguments{
\item{path}{A file path that is passed to the gen function}

\item{variable_config}{A named list specifying fun and assert. Fun is the
name of the function to apply to the input_folder, while assert (optional)
is a character vector with assertions snippets to apply.}

\item{Where}{to write the data. The function creates a folder
with the same name as the variable folder.}
}
\value{
The function returns a character vector of messages with the results
of assertions. The data is written to disk.
}
\description{
A dispatch function that applies the function specified in variable_config
to the file in path. Subsequently, the result is put through several
assertions, both a base set of assertions from base_assertions and possibly
also custom assertions specified in the variable_config. Then, data is
written to the output_folder in RDS format.
}
