% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_file_httr2}
\alias{download_file_httr2}
\title{Download data from url and save to file}
\usage{
download_file_httr2(url, filepath)
}
\arguments{
\item{url}{The url of the data you want to download.}

\item{filepath}{The filepath to where you want to store the data locally.}
}
\description{
This is a simple wrapper around \code{\link[httr2:request]{httr2::request()}} and \code{\link[httr2:req_perform]{httr2::req_perform()}}.
}
