% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_changed_areas}
\alias{gen_changed_areas}
\title{gen_changed_areas}
\usage{
gen_changed_areas(input_folder)
}
\arguments{
\item{input_folder}{}
}
\value{
a list, one sf df for each crossection
}
\description{
Finds the areas and dates where borders have changed in the cshapes dataset.
}
