% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{gen_changed_areas_base}
\alias{gen_changed_areas_base}
\alias{gen_changed_areas}
\title{gen_changed_areas_base}
\usage{
gen_changed_areas_base(cshp)

gen_changed_areas(cshp)
}
\arguments{
\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\value{
a list, one sf df for each crossection
}
\description{
Finds the areas and dates where borders have changed in the cshapes dataset.
}
\section{Functions}{
\itemize{
\item \code{gen_changed_areas()}: Finds the areas and dates where borders have changed in the cshapes dataset. Memoise/caching wrapper.

}}
