% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cmr.R
\name{gen_cmr_sd}
\alias{gen_cmr_sd}
\title{cmr_sd}
\usage{
gen_cmr_sd(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate standard deviation child malnutrition variable.
}
