% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_coastdist}
\alias{gen_coastdist}
\title{gen_coastdist}
\usage{
gen_coastdist(input_folder)
}
\arguments{
\item{input_folder}{path to PRIO-GRID input data}
}
\value{
a tibble, with x, y, pgid, and coastdist
}
\description{
Calculates the distance between cell centroids and the coast.
}
