% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crupet.R
\name{gen_cru_pet}
\alias{gen_cru_pet}
\title{Generate PRIO-GRID Compatible CRU Potential Evapotranspiration Data}
\usage{
gen_cru_pet(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link{pg_current_config}()}.}
}
\value{
A \code{SpatRaster} object with PRIO-GRID spatio-temporal resolution.
}
\description{
This function processes CRU TS v4.09 potential evapotranspiration (PET)
data by aggregating it temporally and spatially to match PRIO-GRID
specifications.
}
\examples{
\dontrun{
r <- gen_cru_pet()
}

}
\seealso{
\code{\link{read_cru_pet}}, \code{\link{pg_date_intervals}},
\code{\link{robust_transformation}}, \code{\link{get_pgfile}}
}
