% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_crutmp.R
\name{gen_cru_tmp}
\alias{gen_cru_tmp}
\title{Generate PRIO-GRID Compatible CRU Near Surface Temperature Data}
\usage{
gen_cru_tmp(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link{pg_current_config}()}.}
}
\value{
A \code{SpatRaster} object (from the \pkg{terra} package) with spatio-temporal
resolution as defined in PRIO-GRID.
}
\description{
This function processes CRU TS (Climate Research Unit Time Series) v4.09
near surface temperature data by aggregating it both temporally and spatially
to match PRIO-GRID specifications. The function takes monthly CRU temperature
data and transforms it to the temporal resolution defined by PRIO-GRID date
intervals (which may be quarterly, yearly, or other intervals), while also
performing spatial aggregation to the PRIO-GRID resolution.
}
\details{
This takes the CRU TS dataset and aggregates it in time and space to
PRIO-GRID specifications.
}
\examples{
\dontrun{
r <- gen_cru_tmp()
}

}
\references{
\insertRef{harrisVersion4CRU2020}{priogrid}
}
\seealso{
\code{\link{read_cru_tmp}} for reading raw CRU temperature data,
\code{\link{pg_date_intervals}} for PRIO-GRID temporal boundaries,
\code{\link{robust_transformation}} for spatial aggregation,
\code{\link{get_pgfile}} for file management
}
