% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{gen_cshp_cover}
\alias{gen_cshp_cover}
\title{gen_cshp_cover}
\usage{
gen_cshp_cover(measurement_date, cshp = read_cshapes())
}
\arguments{
\item{measurement_date}{A single date as Date object}

\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\description{
Takes the cShapes data and returns
a raster-mask that is true for the grid cells that intersects with country borders
included in the international state system at the measurement_date
}
