% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_diamonds_s}
\alias{gen_diamonds_s}
\title{Generate static diamond presence dummy
including records without known start year.}
\usage{
gen_diamonds_s(diamond_data)
}
\arguments{
\item{diamond_data}{DIADATA shapefile from the PRIO Diamond Resources dataset.}
}
\description{
Generate static diamond presence dummy
including records without known start year.
}
