% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_diamprim_s}
\alias{gen_diamprim_s}
\title{Generate static diamond presence dummy
including records without known start year.}
\usage{
gen_diamprim_s(input_folder, variable = "diamprim_s")
}
\description{
Generate static diamond presence dummy
including records without known start year.
}
