% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_drug_y}
\alias{gen_drug_y}
\title{Generate yearly dummy identifying ongoing large-scale drug cultivation
(coca bush, opium poppy and/or cannabis).}
\usage{
gen_drug_y(input_folder)
}
\description{
Generate yearly dummy identifying ongoing large-scale drug cultivation
(coca bush, opium poppy and/or cannabis).
}
