% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_epr_regional}
\alias{gen_epr_regional}
\title{epr_regional}
\usage{
gen_epr_regional(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate variable that counts the number of regionally based ethnic groups
that are settled in a given grid cell in a given year.
}
