% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_excluded}
\alias{gen_excluded}
\title{epr_excluded}
\usage{
gen_excluded(input_folder, variable = "epr_excluded")
}
\arguments{
\item{input_folder}{path to [pg-folder].}

\item{variable}{one of c("epr_excluded", "epr_promoted", "epr_demoted", "epr_regional")}
}
\description{
Counts the number of excluded groups (discriminated or powerless), as defined in the GeoEPR/EPR data
on the status and location of politically relevant ethnic groups, that are settled in the grid cell in a given year.
Group status derived from the EPR Core 2019 dataset, and location derived from the GeoEPR 2019 data. The variable counts only
groups that have a distinct subnational settlement pattern.

Link to original data: https://icr.ethz.ch/data/epr/

Please cite: Vogt, Manuel, Nils-Christian Bormann, Seraina Rüegger, Lars-Erik Cederman, Philipp Hunziker, and Luc Girardin. 2015. “Integrating Data on Ethnicity, Geography, and Conflict: The Ethnic Power Relations Data Set Family.” Journal of Conflict Resolution 59(7): 1327–42.
}
