% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gbd.R
\name{gen_gbd_stunting}
\alias{gen_gbd_stunting}
\title{gbd_stunting}
\usage{
gen_gbd_stunting(input_folder, wasting = FALSE)
}
\arguments{
\item{input_folder}{path to [pg-folder].}
}
\description{
Generate mean stunting
for years 2000 - 2017 using data from
the IHME Low- and Middle-Income Country Child Growth Failure Geospatial Estimates 2000-2017.

Link to original data: http://ghdx.healthdata.org/record/ihme-data/lmic-child-growth-failure-geospatial-estimates-2000-2017.

Please cite: Institute for Health Metrics and Evaluation (IHME). Low- and Middle-Income Country Child Growth Failure Geospatial Estimates 2000-2017. Seattle, United States of America: Institute for Health Metrics and Evaluation (IHME), 2020.
}
