% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nordhaus.R
\name{gen_gcp_mer}
\alias{gen_gcp_mer}
\title{gcp_mer}
\usage{
gen_gcp_mer(
  input_folder,
  variable = "GCP_MER",
  interpolate_time = FALSE,
  interpolate_missing = TRUE
)
}
\arguments{
\item{variable}{one of c("GCP_MER", "GCP_PPP")}

\item{interpolate_time}{if `TRUE`, data is interpolated to all years from 1990 to 2005 See interpolate_pg_timeseries() for details.}

\item{interpolate_missing}{if `TRUE`, interpolates data to grid cells with missing values around coastline.}

\item{gecon_data}{path to [pg-folder].}
}
\description{
Indicates the gross cell product, measured in USD, based on the G-Econ dataset v. 4.0.
The original G-Econ data represent the total economic activity at a 1x1 degree resolution, so when assigning this to PRIO-GRID
we distribute the total value across the number of contained PRIO-GRID land cells. In border areas, the G-Econ 1x1 degree cells
might overlap with PRIO-GRID cells allocated to a neighboring country. To minimize bias, PRIO-GRID only extracts G-Econ data for
cells that have the same country code as the G-Econ cell represents. This variable is only available for five-year intervals between 1990 and 2005,
but function allows for interpolation of yearly values from 1990 to 2005.
}
\details{
Link to original data: https://gecon.yale.edu/.

Please cite: Nordhaus, William D. (2006) Geography and macroeconomics: New data and new findings. Proceedings of the National Academy of Sciences of the USA, 103(10): 3510-3517.
}
