% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nordhaus.R
\name{gen_gcp_ppp}
\alias{gen_gcp_ppp}
\title{Generate gcp_ppp variable}
\usage{
gen_gcp_ppp(input_folder)
}
\description{
Generate gross cell product (USD adjusted for purchasing-power-parity) variable,
based on the G-Econ dataset v. 4.0.
}
