% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_gem_s}
\alias{gen_gem_s}
\title{Generate static gem dummy.}
\usage{
gen_gem_s(input_folder)
}
\description{
Function to generate static gem presence dummy for records without known discovery or start of production year.
}
