% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_gem_y}
\alias{gen_gem_y}
\title{gem_y}
\usage{
gen_gem_y(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
gem deposits have been found within the given grid cell,
based on the GEMDATA dataset. This variable only codes
those deposits that have a known year of discovery or
start of production. For a complete picture, these data
must therefore be combined with the gem_s data.
Available only until 2004.
}
