% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_geoepr}
\alias{gen_geoepr}
\title{Generate excluded variable}
\usage{
gen_geoepr(path)
}
\arguments{
\item{geoepr_data}{GeoEPR 2018 shapefile.}

\item{epr_data}{EPR Core 2018 .csv data.}
}
\description{
Counts the number of excluded groups of politically relevant ethnic groups
settled in a given grid cell in a given year.
}
