% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glcshare.R
\name{gen_glcs_mangroves}
\alias{gen_glcs_mangroves}
\title{gen_glcs_mangroves}
\usage{
gen_glcs_mangroves(input_folder)
}
\arguments{
\item{input_folder}{}
}
\value{
a dataframe with x, y, and share mangroves
}
\description{
Aggregates and transforms the FAO Global Land Cover Share Database (2013) "mangroves" to PRIO-GRID.
}
\details{
Link to original data: http://www.fao.org/geonetwork/srv/en/metadata.show?id=47948

Please cite: John Latham, Renato Cumani, Ilaria Rosati and Mario Bloise (2014) FAO Global Land Cover (GLC-SHARE) Beta-Release 1.0 Database, Land and Water Division, J, available at: http://www.fao.org/uploads/media/glc-share-doc.pdf.
}
