% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glcshare.R
\name{gen_glcshare}
\alias{gen_glcshare}
\title{gen_glcshare}
\usage{
gen_glcshare(input_folder, land_type)
}
\arguments{
\item{land_type}{one of c("artificial_surfaces", "cropland", "grassland", "trees", "shrubs", "herbaceous", "mangroves", "sparse", "bare_soil", "snow", "water")}
}
\value{
a dataframe with x, y, and land type share
}
\description{
Aggregates and transforms the FAO Global Land Cover Share Database (2013) to PRIO-GRID.
}
\details{
Link to original data: http://www.fao.org/geonetwork/srv/en/metadata.show?id=47948

Please cite: John Latham, Renato Cumani, Ilaria Rosati and Mario Bloise (2014) FAO Global Land Cover (GLC-SHARE) Beta-Release 1.0 Database, Land and Water Division, J, available at: http://www.fao.org/uploads/media/glc-share-doc.pdf.
}
