% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gmted2010.R
\name{gen_gmted2010}
\alias{gen_gmted2010}
\title{gen_gmted2010}
\usage{
gen_gmted2010(input_folder)
}
\arguments{
\item{input_folder}{}
}
\value{
a dataframe with x, y, and gmted2010 elevation
}
\description{
Aggregates and transforms the gmted2010 topography data to PRIO-GRID. Source: https://www.usgs.gov/land-resources/eros/coastal-changes-and-impacts/gmted2010
}
