% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldsurface_s}
\alias{gen_goldsurface_s}
\title{goldsurface_s}
\usage{
gen_goldsurface_s(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
surface gold deposits have been found within the given grid cell,
based on the dGOLD_S subset of the GOLDATA v1.2 dataset.
This variable only codes those deposits that do not have a known
year of discovery or start of production. For a complete picture,
these data must therefore be combined with the goldsurface_y data.
}
