% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldsurface_y}
\alias{gen_goldsurface_y}
\title{Generate yearly dummy identifying semi-lootable surface gold deposits.}
\usage{
gen_goldsurface_y(input_folder)
}
\description{
Generate yearly dummy identifying semi-lootable surface gold deposits.
}
