% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldvein_s}
\alias{gen_goldvein_s}
\title{goldvein_s}
\usage{
gen_goldvein_s(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
vein gold deposits have been found within the given grid cell,
based on the dGOLD_NL subset of the GOLDATA v1.2 dataset.
This variable only codes those deposits that do not have a known
year of discovery or start of production. For a complete picture,
these data must therefore be combined with the goldvein_y data.
}
