% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldvein_y}
\alias{gen_goldvein_y}
\title{Generate yearly dummy identifying non-lootable vein gold deposits.}
\usage{
gen_goldvein_y(input_folder)
}
\description{
Generate yearly dummy identifying non-lootable vein gold deposits.
}
