% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mirca.R
\name{gen_grow_start_end}
\alias{gen_grow_start_end}
\title{Generate grow_start and grow_end variables}
\usage{
gen_grow_start_end(mirca, no_months = 5)
}
\arguments{
\item{mirca}{MIRCA2000 data (.TXT).}

\item{no_months}{Width of rolling window (number of months) used to compute maximum sum of harvested area. Defaults to 5.}
}
\value{
PRIO-GRID-consistend RasterBrick with one layer for each variable.
}
\description{
Function to generate grow_start and grow_end 
based on MIRCA2000 data.
}
