% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mirca.R
\name{gen_growarea}
\alias{gen_growarea}
\title{grow_area}
\usage{
gen_growarea(input_folder, month_no = FALSE)
}
\arguments{
\item{input_folder}{path to [pg-folder].}

\item{month_no}{if `TRUE`, months are reported as numeric, 1-12.}
}
\value{
a dataframe with x, y, pgid, month, and grow_area proportion.
}
\description{
Monthly variable, providing the area proportion of each cell
that is being harvested each month. Based on the Cropping Periods List data
from the MIRCA2000 dataset v.1.1. Shapshot for the year 2000 only.
}
\details{
Link to original data: https://www.uni-frankfurt.de/45218031

Please cite: Portmann, F. T., Siebert, S. & Döll, P. (2010). MIRCA2000 – Global monthly irrigated and rainfed crop areas around the year 2000: A new high-resolution data set for agricultural and hydrological modeling, Global Biogeochemical Cycles, 24, GB 1011, doi:10.1029/2008GB003435.
}
