% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mirca.R
\name{gen_growend}
\alias{gen_growend}
\title{grow_end}
\usage{
gen_growend(input_folder, no_months = 5)
}
\arguments{
\item{input_folder}{path to [pg-folder].}

\item{no_months}{width of rolling window (number of months) used to compute max sum of harvested area. Defaults to 5.}

\item{variable}{one of c("grow_start", "grow_end")}
}
\value{
a dataframe with x, y, pgid, and grow_end month.
}
\description{
Provides the final month of the growing season in the cell, based on the Cropping Periods List data
from the MIRCA2000 dataset v.1.1. Snapshot for the year 2000 only.
Values range from 1-12. Defined as the last of five consecutive months with the highest harvested area.
}
\details{
Link to original data: https://www.uni-frankfurt.de/45218031

Please cite: Portmann, F. T., Siebert, S. & Döll, P. (2010). MIRCA2000 – Global monthly irrigated and rainfed crop areas around the year 2000: A new high-resolution data set for agricultural and hydrological modeling, Global Biogeochemical Cycles, 24, GB 1011, doi:10.1029/2008GB003435.
}
