% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mirca.R
\name{gen_growseas}
\alias{gen_growseas}
\title{Generate monthly growseas variable}
\usage{
gen_growseas(mirca)
}
\arguments{
\item{mirca}{MIRCA2000 data (.TXT)}
}
\description{
Function to generate the area proportion of each cell
that is being harvested for each month.
}
