% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_gwcode}
\alias{gen_gwcode}
\title{gen_gwcode}
\usage{
gen_gwcode(
  input_folder,
  input_file = "gwcode_changes.parquet",
  interval = "1 year",
  enddate = as.Date("2020-1-1"),
  startdate = NULL
)
}
\arguments{
\item{interval}{see by argument in base::seq.Date. default is 1 year.}

\item{enddate}{a date. option to add more years than the last entry in the input data}

\item{startdate}{a date. filters the input data to start at the given date.}
}
\description{
Takes gwcode_changes and translates it into a regular time-series.
}
