% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_gwcode_changes}
\alias{gen_gwcode_changes}
\title{gen_gwcode_changes}
\usage{
gen_gwcode_changes(input_folder)
}
\arguments{
\item{fname}{File path to Weidmann cshapes data}
}
\description{
Takes the weidmann cshapes data set and returns a
rasterstack that encodes the country ownership of a cell as it looked like on the 1st each month.
The rasterstack only includes the months where there was a change from one month to the next.
}
