% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_gwcode_month}
\alias{gen_gwcode_month}
\title{gen_gwcode_month}
\usage{
gen_gwcode_month(fname, numCores = 1, quiet = quiet,
  output_folder = NULL)
}
\arguments{
\item{fname}{File path to Weidmann cshapes data}

\item{numCores}{Number of cores to use in calculation. Windows-users can only use 1. Using parallel::mclapply.}

\item{quiet}{Whether or not sf::st_ functions should print warnings.}
}
\description{
Takes the weidmann cshapes data set and returns a
rasterstack that encodes the country ownership of a cell as it looked like on the 1st each month.
The rasterstack only includes the months where there was a change from one month to the next.
}
