% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imr_ihme.R
\name{gen_imr_ihme_mean}
\alias{gen_imr_ihme_mean}
\title{imr_ihme_mean

Generate variable of average under-5 mortality rate within each grid cell
based on Institute for Health Metrics and Evaluation (IHME) data (available only for Africa).}
\usage{
gen_imr_ihme_mean(input_folder, fun = "mean")
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
imr_ihme_mean

Generate variable of average under-5 mortality rate within each grid cell
based on Institute for Health Metrics and Evaluation (IHME) data (available only for Africa).
}
