% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imr_ihme.R
\name{gen_imr_ihme_sd}
\alias{gen_imr_ihme_sd}
\title{imr_ihme_sd

Generate standard deviation under-5 mortality rate variable within each grid cell
based on Institute for Health Metrics and Evaluation (IHME) data (available only for Africa).}
\usage{
gen_imr_ihme_sd(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
imr_ihme_sd

Generate standard deviation under-5 mortality rate variable within each grid cell
based on Institute for Health Metrics and Evaluation (IHME) data (available only for Africa).
}
