% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{gen_land}
\alias{gen_land}
\title{gen_pgland}
\usage{
gen_land()
}
\description{
Takes the NaturalEarth 50m raster and returns
a raster-mask that is true for the grid cells that intersects with land.
}
