% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{gen_lifexp}
\alias{gen_lifexp}
\title{Generate PRIO-GRID Life Expectancy}
\usage{
gen_lifexp(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link{pg_current_config}()}.}
}
\value{
A \code{SpatRaster} object containing PRIO-GRID life expectancy values.
}
\description{
Convenience wrapper around \code{\link{shdi}} that generates a PRIO-GRID–aligned
raster of life expectancy at birth.
}
\examples{
\dontrun{
lifexp_pg <- gen_lifexp()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{shdi}}
}
