% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{gen_ne_cover}
\alias{gen_ne_cover}
\title{gen_ne_cover}
\usage{
gen_ne_cover()
}
\description{
Takes the NaturalEarth 50m raster and returns
a raster-mask that is true for the grid cells that intersects with land.
}
