% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{gen_ne_coversh}
\alias{gen_ne_coversh}
\title{gen_ne_coversh}
\usage{
gen_ne_coversh()
}
\description{
Takes the NaturalEarth 10m raster and returns a raster
in PRIO-GRID resolution with the share of the cell that intersects with land.
}
