% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nightlights.R
\name{gen_nightlights}
\alias{gen_nightlights}
\title{nightlights}
\usage{
gen_nightlights(input_folder)
}
\arguments{
\item{input_folder}{path to [pg-folder].}
}
\description{
Generate yearly nighttime light variable, available from 1992-2018.
Please cite: Li, X., Zhou, Y., Zhao, M. et al. A harmonized global nighttime light dataset 1992–2018. Sci Data 7, 168 (2020). https://doi.org/10.1038/s41597-020-0510-y
}
