% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_petro_s}
\alias{gen_petro_s}
\title{Generate static petroleum presence dummy variable
for records without known year of discovery or production start.}
\usage{
gen_petro_s(petro_data)
}
\arguments{
\item{petro_data}{PRIO Petroleum Dataset v. 1.2 onshore shapefile}
}
\description{
Generate static petroleum presence dummy variable
for records without known year of discovery or production start.
}
