% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_petro_y}
\alias{gen_petro_y}
\title{Generate yearly petroleum presence dummy variable.}
\usage{
gen_petro_y(petro_data)
}
\arguments{
\item{petro_data}{PRIO Petroleum Dataset v. 1.2 onshore shapefile}
}
\description{
Generate yearly petroleum presence dummy variable.
}
