% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_petroleum_s}
\alias{gen_petroleum_s}
\title{petroleum_s}
\usage{
gen_petroleum_s(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
onshore petroleum deposits have been found within
the given grid cell, based on the Petroleum Dataset v1.2.
This variable only codes those petroleum deposits that do not
have a known year of discovery or start of production.
For a complete picture, these data must therefore be
combined with the petroleum_y data.

Link to original data: https://www.prio.org/Data/Geographical-and-Resource-Datasets/Petroleum-Dataset/Petroleum-Dataset-v-12/

Please cite: Lujala, Päivi; Jan Ketil Rød & Nadia Thieme, 2007. ‘Fighting over Oil: Introducing A New Dataset’, Conflict Management and Peace Science 24(3), 239-256.
}
