% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_petroleum_y}
\alias{gen_petroleum_y}
\title{petroleum_y}
\usage{
gen_petroleum_y(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
onshore petroleum deposits have been found within
the given grid cell for any given year, based on the
Petroleum Dataset v1.2. This variable only codes those
petroleum deposits that have a known year of discovery or
start of production. For a complete picture, these data
must therefore be combined with the petroleum_s data.
Available only until 2003.
}
