% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gpw.R
\name{gen_pop_gpw}
\alias{gen_pop_gpw}
\title{Generate pop_gpw_ variables.}
\usage{
gen_pop_gpw(gpw_nc, fun)
}
\arguments{
\item{gpw_nc}{Gridded Population of the World v.4 NetCDF file.}

\item{fun}{Function to aggregate by ('sum', 'sd', 'min', 'max')}
}
\description{
Generate pop_gpw_ variables.
}
