% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gpw.R
\name{gen_pop_gpw_max}
\alias{gen_pop_gpw_max}
\title{pop_gpw_max}
\usage{
gen_pop_gpw_max(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate maximum population count
for years 2000, 2005, 2010, 2015 and 2020 from
the Gridded Population of the World v.4 data.
}
