% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_promoted}
\alias{gen_promoted}
\title{epr_promoted}
\usage{
gen_promoted(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate variable that counts the number of politically relevant ethnic groups
that are settled in a given grid cell in a given year, and that have increased their political access
from the previous year.
}
