% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_riverdist}
\alias{gen_riverdist}
\title{gen_riverdist}
\usage{
gen_riverdist(input_folder)
}
\arguments{
\item{input_folder}{path to PRIO-GRID input data}
}
\value{
a tibble, with x, y, pgid, and riverdist
}
\description{
Calculates the distance between cell centroids and the nearest major river or lake.
}
