% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ruggedness.r
\name{gen_ruggedness}
\alias{gen_ruggedness}
\title{gen_ruggedness}
\usage{
gen_ruggedness(input_folder, interpolate_missing = FALSE)
}
\arguments{
\item{input_folder}{path to [pg-folder].}

\item{interpolate_missing}{if `TRUE`, interpolates data to grid cells with missing values around coastline.}
}
\value{
a dataframe with x, y, and ruggedness
}
\description{
Aggregates and transforms the ruggedness data to PRIO-GRID.
}
