% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ttime.R
\name{gen_ttime_max}
\alias{gen_ttime_max}
\title{ttime_max}
\usage{
gen_ttime_max(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate maximum travel time to
nearest urban center within each cell.
Snapshot for 2014 only. Drawn from the
Malaria Atlas Project's Accessibility to Cities data.
}
