% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ucdp.R
\name{gen_ucdpged_distance_within_country}
\alias{gen_ucdpged_distance_within_country}
\title{Generate UCDP GED distance within country}
\usage{
gen_ucdpged_distance_within_country(
  ged = read_ucdp_ged(),
  cshp = read_cshapes()
)
}
\arguments{
\item{ged}{UCDP GED sf data frame, defaults to \code{\link[=read_ucdp_ged]{read_ucdp_ged()}}}

\item{cshp}{cShapes sf data frame, defaults to \code{\link[=read_cshapes]{read_cshapes()}}}
}
\value{
SpatRaster with multiple layers, one for each time interval in \code{\link[=pg_date_intervals]{pg_date_intervals()}}.
}
\description{
Calculates UCDP GED distance to the nearest UCDP GED even within each country for each time interval in \code{\link[=pg_date_intervals]{pg_date_intervals()}}.
}
