% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_waterrisk.R
\name{gen_waterrisk}
\alias{gen_waterrisk}
\title{waterrisk}
\usage{
gen_waterrisk(input_folder)
}
\arguments{
\item{input_folder}{path to [pg-folder].}
}
\description{
Generate variable measuring overall waterrisk, drawn from the yearly Aqueduct 3.0 dataset.
Please cite: Hofste, R., S. Kuzma, S. Walker, E.H., Sutanudjaja, et. al. 2019. “Aqueduct 3.0: Updated Decision-Relevant Global Water Risk Indicators.” Technical Note. Washington, DC: World Resources Institute. Available online at: https://www.wri.org/publication/aqueduct-30.
}
