% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{get_bib_element}
\alias{get_bib_element}
\title{Extract bibliography element}
\usage{
get_bib_element(key, element = "author", as_character = TRUE)
}
\arguments{
\item{key}{A bibkey found in \link{pgsources}}

\item{element}{Supports author, journal, year, and title.}

\item{as_character}{Return the result as a string instead of a RefManageR class object.}
}
\value{
BibEntry-element class or vector with character strings
}
\description{
Get bibliography information. Is used by \code{\link[=pgsearch]{pgsearch()}} to search \link{pgsources} based
on bibliography information.
}
\examples{
get_bib_element("schvitzMappingInternationalSystem2022", element = "author")
}
