% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{get_bibliography}
\alias{get_bibliography}
\title{Get bibliography given a bibkey}
\usage{
get_bibliography(keys = NULL, as_biblatex = FALSE)
}
\arguments{
\item{as_biblatex}{}
}
\description{
Get bibliography given a bibkey
}
