% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_closest_distance}
\alias{get_closest_distance}
\title{get_closest_distance}
\usage{
get_closest_distance(points, features, check_dateline = TRUE)
}
\arguments{
\item{points}{point data of class sfg, sfc, or sf}

\item{features}{object of class sfg, sfc, or sf}

\item{check_dateline}{boolean, whether or not to account for the possibility that the shortest distance is crossing the dateline. doubles the time it takes.}
}
\value{
A tibble with columns x, y, and value.
}
\description{
Finds (element-wise) the closest distance between a set of points and a set of features.
}
