% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity.R
\name{get_folder_fingerprint}
\alias{get_folder_fingerprint}
\title{get_folder_fingerprint}
\usage{
get_folder_fingerprint(root)
}
\description{
Gets a hash fingerprint, which can be used to ascertain that
the contents of a folder are intact.
}
