% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_format_spec}
\alias{get_format_spec}
\title{get_format_spec}
\usage{
get_format_spec(format_name)
}
\description{
Returns the spec for a format, with functions for reading and
writing as well as the file extension to use.
Used for dispatching gen functions with different file formats.
See extdata/format_specification.yaml for further information.
}
