% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imr_ihme.R
\name{get_ihme_brick}
\alias{get_ihme_brick}
\title{get_ihme_brick}
\usage{
get_ihme_brick(folder)
}
\arguments{
\item{folder}{A folder containing the files}
}
\description{
Returns a raster brick containing the IHME data
}
