% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imr_ihme.R
\name{get_ihme_files}
\alias{get_ihme_files}
\title{get_ihme_files}
\usage{
get_ihme_files(input_folder)
}
\arguments{
\item{folder}{A folder containing the files}
}
\value{
A character vector with four paths
}
\description{
Returns the paths to the four files needed by the functions
generating the ihme raster
}
