% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{getNeighbors}
\alias{getNeighbors}
\title{Get nth order neighbors in PRIO-GRID.
NB! This does not work!}
\usage{
getNeighbors(gid, norder = 1, ncol = 720, nrow = 360,
  includeSelf = FALSE)
}
\arguments{
\item{gid}{The GID to the cell that you would like to find the neighbors of.}

\item{norder}{The nth order of neighbors.}

\item{ncol}{The number of columns in the grid.}

\item{nrow}{The number of rows in the grid.}

\item{includeSelf}{Whether self should be part of the result or not.}
}
\value{
A vector with gid's of neighbors.
}
\description{
Get nth order neighbors in PRIO-GRID.
NB! This does not work!
}
\examples{
pg <- create_pg_indices(3, 3)
getNeighbors(5, 1, ncol=3, nrow=3)
}
