% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{getXY}
\alias{getXY}
\title{Get column and row number based on PRIO-GRID ID with ncol and nrow.}
\usage{
getXY(gid, ncol, nrow)
}
\arguments{
\item{gid}{An integer cell id.}

\item{ncol}{The number of columns in the grid.}

\item{nrow}{The number of rows in the grid.}
}
\value{
A vector with components
\itemize{
  \item X Column number.
  \item Y Row number.
}
}
\description{
Get column and row number based on PRIO-GRID ID with ncol and nrow.
}
\examples{
pg <- create_pg_indices(3, 3)
getXY(5, 3, 3)
}
