% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{interpolate_crossection}
\alias{interpolate_crossection}
\title{interpolate_crossection}
\usage{
interpolate_crossection(crossection, variable, lon, lat)
}
\arguments{
\item{crossection}{a dataframe crossection lon, lat, variable}

\item{variable}{the variable to interpolate missing from}

\item{lon}{a string denoting the name of the longitude variable}

\item{lat}{a string denoting the name of the latitude variable}
}
\value{
the interpolated crossection
}
\description{
Increasingly coarse bilinear interpolation of missing data
}
