% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{interpolate_pg_timeseries}
\alias{interpolate_pg_timeseries}
\title{interpolate_pg_timeseries}
\usage{
interpolate_pg_timeseries(
  pgdf,
  variable,
  date_var = "year",
  interval = "1 year",
  startdate = NULL,
  enddate = NULL,
  add_pg_index = TRUE
)
}
\arguments{
\item{pgdf}{output from PRIO-GRID gen_*() function.}

\item{variable}{the variable to interpolate missing values from.}

\item{date_var}{string denoting the name of the date variable. Defaults to "year".}

\item{interval}{string denoting the interval between interpolated values, i.e. the temporal resolution. Defaults to "1 year".}

\item{startdate}{starting date. Defaults to the earliest observation in the original data.}

\item{enddate}{end date. Defaults to the latest observation in the original data.}

\item{add_pg_index}{boolean, whether to add a pgid column or not. Default is TRUE.}
}
\value{
Interpolated time-series data frame.
}
\description{
Interpolation of time-series data where only some cross-sections are available.
}
\details{
`interval` can be specified as `"day"`, `"week"`, `"month"`, `"quarter"` or `"year"`. See `seq.Date` for details.
}
