% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gwcode.R
\name{landgrid}
\alias{landgrid}
\title{Landgrid}
\usage{
landgrid(cshapes, blank = NULL)
}
\arguments{
\item{cshapes}{An sf object with cshapes}
}
\description{
Generate a raster that only contains gridcells on 
or close to landmasses. Hopefully improves performance 
of some functions.
}
