% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priogrid.R
\name{make_pg}
\alias{make_pg}
\title{Make PRIOGrid from raw data}
\usage{
make_pg(input_folder, output_folder)
}
\description{
This is the "master function" that turns raw data
into PRIOGrid.
Not functional, only for illustration.
 @param input_folder Location of folder containing 
folders of raw data
 @param output_folder Output folder
}
