% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{missing_in_pg}
\alias{missing_in_pg}
\title{missing_in_pg}
\usage{
missing_in_pg(df, plot_missing = TRUE, ...)
}
\arguments{
\item{df}{a dataframe crossection x, y, variable}

\item{...}{function accepts other parameters passed to raster::plot()}

\item{variable}{the variable to convert to plot}
}
\value{
A dataframe with the cells that are missing and a plot.
}
\description{
Finds whether a crossection has missing data for any of the 64818 PRIO-GRID land cells,
and plots them.
}
