% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{nth_order_pgneighbors}
\alias{nth_order_pgneighbors}
\title{Get nth-order neighbors in PRIO-GRID, vectorized version.}
\usage{
nth_order_pgneighbors(gid, order, ncol, nrow, include_self = FALSE)
}
\arguments{
\item{gid}{The GIDs to the cells that you would like to find the neighbors of.}

\item{order}{A number, the n-th order.}

\item{ncol}{The number of columns in the grid.}

\item{nrow}{The number of rows in the grid.}

\item{include_self}{Whether or not to also return the center gids}
}
\value{
A dataframe with a list of neighbors for each gid in gids.
}
\description{
Get nth-order neighbors in PRIO-GRID, vectorized version.
}
