% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{panel_to_pg}
\alias{panel_to_pg}
\title{panel_to_pg}
\usage{
panel_to_pg(df, timevar, variable, need_aggregation, missval, fun)
}
\arguments{
\item{df}{a sf (simple features) data frame with panel data}

\item{timevar}{the temporal variable (used to split data into crossections)}

\item{variable}{the variable to convert to PRIO-GRID}

\item{need_aggregation}{if FALSE, will use velox, which only extracts the last value for each feature. fast when applicable.}

\item{missval}{only used if need_aggregation is FALSE. velox sets missing data to this value.}

\item{fun}{aggregation function. only used if need_aggregation is TRUE.}
}
\value{
A tibble with x, y, pgid, timevar, and variable.
}
\description{
Converts a sf data frame with panel data to PRIO-GRID.
}
