% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pg_current_config}
\alias{pg_current_config}
\title{Get the current session configuration}
\usage{
pg_current_config()
}
\value{
A \code{pg_config} object.
}
\description{
Returns the current session-scoped PRIO-GRID config. If none has been set,
initializes with defaults from \code{\link[=pg_config]{pg_config()}}.
}
