% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pg_date_intervals}
\alias{pg_date_intervals}
\title{Get a sequence of date intervals}
\usage{
pg_date_intervals(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link[=pg_current_config]{pg_current_config()}}.}
}
\value{
Date interval vector
}
\description{
The start date is the last observed date in the first interval of the intervals you
want to capture. E.g., if start date is 31 January 2010 with 1 month resolution, then the
first interval goes from 1 January 2010 to 31 January 2010. The intervals are not necessary
equal length.
}
\examples{
pg_date_intervals()
}
