% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pg_dates}
\alias{pg_dates}
\title{Get a sequence of dates}
\usage{
pg_dates(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link[=pg_current_config]{pg_current_config()}}.}
}
\value{
Date vector
}
\description{
This is a wrapper of \link[base:seq.Date]{base::seq.Date}, only with defaults that can be set in the config.
}
\examples{
pg_dates()
}
