% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{pg_rawfiles}
\alias{pg_rawfiles}
\title{Extract url- and file-info from PRIO-GRID metadata}
\usage{
pg_rawfiles(use_mirror = TRUE, only_file_extensions = FALSE)
}
\arguments{
\item{use_mirror}{Boolean. Whether or not to use PRIO-GRID mirror.}

\item{only_file_extensions}{Logical. If TRUE, returns file extensions only. Used for testing. Default FALSE.}
}
\value{
data.frame
}
\description{
Extract url- and file-info from PRIO-GRID metadata
}
\examples{
file_info <- pg_rawfiles()
}
