% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pg_rawfolder}
\alias{pg_rawfolder}
\title{Get the raw data folder path}
\usage{
pg_rawfolder()
}
\value{
Character string with the raw data folder path.
}
\description{
Returns the persistent rawfolder path. Errors with a helpful message if unset.
}
