% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pg_reset_config}
\alias{pg_reset_config}
\title{Reset session configuration to defaults}
\usage{
pg_reset_config()
}
\value{
The default \code{pg_config} object (invisibly).
}
\description{
Resets the current session config to the defaults from \code{\link[=pg_config]{pg_config()}}.
}
