% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{pg_response_headers}
\alias{pg_response_headers}
\title{Parse the headers of one row of a curl download report}
\usage{
pg_response_headers(report, i)
}
\arguments{
\item{report}{data.frame from \code{\link[curl:multi_download]{curl::multi_download()}}.}

\item{i}{Integer. Row number.}
}
\value{
Named list of headers, empty if unavailable.
}
\description{
Parse the headers of one row of a curl download report
}
\keyword{internal}
