% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pg_temporal_unit}
\alias{pg_temporal_unit}
\title{Temporal unit for a config's resolution}
\usage{
pg_temporal_unit(config = pg_current_config())
}
\description{
Temporal unit for a config's resolution
}
\keyword{internal}
