% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{pg_timeseries_from_changes}
\alias{pg_timeseries_from_changes}
\title{pg_timeseries_from_changes}
\usage{
pg_timeseries_from_changes(changes_df, interval, enddate, startdate)
}
\arguments{
\item{interval}{see by argument in base::seq.Date}
}
\description{
#' Takes the weidmann cshapes data set and returns a
rasterstack for each year that codes the country ownership of each cell
as it looked like 1 January each year.
}
