% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgchecksum}
\alias{pgchecksum}
\title{PRIO-GRID File Checksums (MD5)}
\format{
\subsection{\code{pgchecksum}}{

A data frame with 22 rows and 5 columns:
\describe{
\item{source_name}{Full name of the source, preferably including the institution hosting it.}
\item{source_version}{The version of the data source, as noted by the creator. If none, use the publication year.}
\item{id}{A Unique Universal Identifier (UUID)}
\item{filename}{The filename of a file from the data source.}
\item{md5}{The MD5 checksum of the correct data to use.}
}
}
}
\usage{
pgchecksum
}
\description{
We provide here file checksums for all the files we have used to test and build PRIO-GRID.
You can test if you have managed to download the same files using \code{\link[=check_pgsourcefiles]{check_pgsourcefiles()}}
after you have downloaded the files using \code{\link[=download_pg_rawdata]{download_pg_rawdata()}}.
}
\keyword{datasets}
