% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{pgcitations}
\alias{pgcitations}
\title{Print citations for PRIO-GRID variables}
\usage{
pgcitations(df, as_biblatex = FALSE)
}
\arguments{
\item{df}{A data.frame or character vector containing variables listed in \link{pgvariables}}

\item{as_biblatex}{Defaults to FALSE. Prints biblatex entry of citations.}
}
\value{
Pure text or biblatex print of citation
}
\description{
Takes the variable names in a data.frame and links with citation using
\link{pgvariables}, \link{pgsources}, and \code{\link[=get_bibliography]{get_bibliography()}}.
}
\examples{
df <- data.frame("hilde_urban" = 1, "cru_tmp" = 1, "geoepr_reg_excluded" = 1)
pgcitations(df)
}
