% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{pgland}
\alias{pgland}
\title{pgland}
\usage{
pgland(input_folder)
}
\arguments{
\item{input_folder}{path to PRIO-GRID input data}
}
\description{
Takes the weidmann cshapes data set and returns a raster for the
grid cells that intersects with land.
}
