% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{pgneighbors}
\alias{pgneighbors}
\title{Get first order neighbors in PRIO-GRID.
TODO: does this work as cell2nb in spdep? Must probably map to PG indices?}
\usage{
pgneighbors(x, ncol, nrow, asmat = TRUE)
}
\arguments{
\item{x}{The GID to the cell that you would like to find the neighbors of.}

\item{ncol}{The number of columns in the grid.}

\item{nrow}{The number of rows in the grid.}
}
\value{
If asmat=TRUE: A matrix of the neighbors, If asmat=FALSE: A vector of the neighbors.
}
\description{
Get first order neighbors in PRIO-GRID.
TODO: does this work as cell2nb in spdep? Must probably map to PG indices?
}
\examples{
pg <- create_pg_indices(5, 5)
pgneighbors(1, ncol=5, nrow=5)
}
