% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R, R/zzz.R
\docType{data}
\name{PGOptionsManager}
\alias{PGOptionsManager}
\alias{pgoptions}
\title{R6 Class to set and get package-wide options for PRIO-GRID.}
\format{
An object of class \code{PGOptionsManager} (inherits from \code{R6}) of length 25.
}
\usage{
pgoptions
}
\description{
Options are stored in a persistent file in AppData, see pgconfig_cache$info()
}
\examples{
pgoptions <- PGOptionsManager$new()
pgoptions$set_rawfolder <- tempfolder()
pgoptions$set_verbose(FALSE)
pgoptions$print()
}
\keyword{datasets}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PGOptionsManager-new}{\code{PGOptionsManager$new()}}
\item \href{#method-PGOptionsManager-reset_options}{\code{PGOptionsManager$reset_options()}}
\item \href{#method-PGOptionsManager-set_extent}{\code{PGOptionsManager$set_extent()}}
\item \href{#method-PGOptionsManager-set_crs}{\code{PGOptionsManager$set_crs()}}
\item \href{#method-PGOptionsManager-set_nrow}{\code{PGOptionsManager$set_nrow()}}
\item \href{#method-PGOptionsManager-set_ncol}{\code{PGOptionsManager$set_ncol()}}
\item \href{#method-PGOptionsManager-set_rawfolder}{\code{PGOptionsManager$set_rawfolder()}}
\item \href{#method-PGOptionsManager-set_verbose}{\code{PGOptionsManager$set_verbose()}}
\item \href{#method-PGOptionsManager-set_automatic_download}{\code{PGOptionsManager$set_automatic_download()}}
\item \href{#method-PGOptionsManager-set_temporal_resolution}{\code{PGOptionsManager$set_temporal_resolution()}}
\item \href{#method-PGOptionsManager-set_start_date}{\code{PGOptionsManager$set_start_date()}}
\item \href{#method-PGOptionsManager-set_end_date}{\code{PGOptionsManager$set_end_date()}}
\item \href{#method-PGOptionsManager-get_crs}{\code{PGOptionsManager$get_crs()}}
\item \href{#method-PGOptionsManager-get_extent}{\code{PGOptionsManager$get_extent()}}
\item \href{#method-PGOptionsManager-get_nrow}{\code{PGOptionsManager$get_nrow()}}
\item \href{#method-PGOptionsManager-get_ncol}{\code{PGOptionsManager$get_ncol()}}
\item \href{#method-PGOptionsManager-get_rawfolder}{\code{PGOptionsManager$get_rawfolder()}}
\item \href{#method-PGOptionsManager-get_verbose}{\code{PGOptionsManager$get_verbose()}}
\item \href{#method-PGOptionsManager-get_automatic_download}{\code{PGOptionsManager$get_automatic_download()}}
\item \href{#method-PGOptionsManager-get_temporal_resolution}{\code{PGOptionsManager$get_temporal_resolution()}}
\item \href{#method-PGOptionsManager-get_start_date}{\code{PGOptionsManager$get_start_date()}}
\item \href{#method-PGOptionsManager-get_end_date}{\code{PGOptionsManager$get_end_date()}}
\item \href{#method-PGOptionsManager-print}{\code{PGOptionsManager$print()}}
\item \href{#method-PGOptionsManager-clone}{\code{PGOptionsManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-new}{}}}
\subsection{Method \code{new()}}{
Initialize options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-reset_options"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-reset_options}{}}}
\subsection{Method \code{reset_options()}}{
Reset options to defaults.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$reset_options()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_extent"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_extent}{}}}
\subsection{Method \code{set_extent()}}{
Set output spatial extent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_extent(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector with extent, c(xmin, xmax, ymin, ymax)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_crs"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_crs}{}}}
\subsection{Method \code{set_crs()}}{
Set output CRS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_crs(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{CRS string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_nrow}{}}}
\subsection{Method \code{set_nrow()}}{
Set number of rows in output raster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_nrow(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Integer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_ncol"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_ncol}{}}}
\subsection{Method \code{set_ncol()}}{
Set number of cols in output raster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_ncol(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Integer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_rawfolder"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_rawfolder}{}}}
\subsection{Method \code{set_rawfolder()}}{
Set the folder where raw-data should be downloaded (possibly very large files), defaults to temporary folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_rawfolder(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{String path to folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_verbose}{}}}
\subsection{Method \code{set_verbose()}}{
Set verbose output of functions if true.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_verbose(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Boolean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_automatic_download"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_automatic_download}{}}}
\subsection{Method \code{set_automatic_download()}}{
Set to true if data downloads should be done automatically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_automatic_download(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Boolean If TRUE, PRIOGRID will attempt to download files that are needed when trying to read specific sources.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_temporal_resolution"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_temporal_resolution}{}}}
\subsection{Method \code{set_temporal_resolution()}}{
Set temporal resolution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_temporal_resolution(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{String increment of temporal sequence. See \link[base:seq.Date]{base::seq.Date} for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_start_date"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_start_date}{}}}
\subsection{Method \code{set_start_date()}}{
Set start date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_start_date(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The date used globally to build PRIO-GRID. The month and day are used to define the
measurement date within the temporal resolution (e.g., as.Date(1900-06-30) would slice June 30 every year for compatible data sources).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-set_end_date"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-set_end_date}{}}}
\subsection{Method \code{set_end_date()}}{
Set end date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$set_end_date(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{a date or "today"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_crs"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_crs}{}}}
\subsection{Method \code{get_crs()}}{
Get crs option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_crs(unparsed = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unparsed}}{Boolean, if TRUE, then return the user input string, else parse with \link[terra:crs]{terra::crs}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_extent"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_extent}{}}}
\subsection{Method \code{get_extent()}}{
Get extent option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_extent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_nrow}{}}}
\subsection{Method \code{get_nrow()}}{
Get nrow option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_nrow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_ncol"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_ncol}{}}}
\subsection{Method \code{get_ncol()}}{
Get ncol option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_ncol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_rawfolder"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_rawfolder}{}}}
\subsection{Method \code{get_rawfolder()}}{
Get rawfolder option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_rawfolder()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_verbose"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_verbose}{}}}
\subsection{Method \code{get_verbose()}}{
Get verbose option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_verbose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_automatic_download"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_automatic_download}{}}}
\subsection{Method \code{get_automatic_download()}}{
Get download prompt option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_automatic_download()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_temporal_resolution"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_temporal_resolution}{}}}
\subsection{Method \code{get_temporal_resolution()}}{
Get temporal resolution option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_temporal_resolution()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_start_date"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_start_date}{}}}
\subsection{Method \code{get_start_date()}}{
Get start date option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_start_date()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-get_end_date"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-get_end_date}{}}}
\subsection{Method \code{get_end_date()}}{
Get end date option
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$get_end_date()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-print"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-print}{}}}
\subsection{Method \code{print()}}{
Prints all options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGOptionsManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PGOptionsManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGOptionsManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
