% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{pgsearch}
\alias{pgsearch}
\title{Search PRIO-GRID meta-data}
\usage{
pgsearch(search_string, bib_element = NULL)
}
\arguments{
\item{search_string}{A character string to search in source name, source version, source id, source tags, spatial extent, temporal
resolution, or in bibliography elements.}

\item{bib_element}{Supports author, journal, year, or title. If null, the search
will not search bibliography elements.}
}
\value{
list with data.frames
}
\description{
Use regex to search the meta-data for the data you are interested in.
}
\examples{
pgsearch("GHSL")
}
