% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{prep_drugs}
\alias{prep_drugs}
\title{Generate drug dummy.}
\usage{
prep_drugs(cannabis_data, coca_data, opium_data)
}
\arguments{
\item{cannabis_data}{Cannabis shapefile from the DRUGDATA dataset.}

\item{coca_data}{Coca bush shapefile from the DRUGDATA dataset.}

\item{opium_data}{Opium poppy shapefile from the DRUGDATA dataset.}
}
\description{
Generate yearly dummy identifying ongoing large-scale drug cultivation
(coca bush, opium poppy and/or cannabis).
}
