% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.pg_config}
\alias{print.pg_config}
\title{Print a pg_config object}
\usage{
\method{print}{pg_config}(x, ...)
}
\arguments{
\item{x}{A \code{pg_config} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a pg_config object
}
