% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gwcode.R
\name{prio_america_eurasia}
\alias{prio_america_eurasia}
\title{America / Eurasia raster}
\usage{
prio_america_eurasia(blank = NULL)
}
\value{
An america / eurasia raster
}
\description{
Returns a blank raster dummy splitting the world into two hemispheres.
}
