% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prio_blank_grid}
\alias{prio_blank_grid}
\title{Create a raster with PRIO-GRID ids.}
\usage{
prio_blank_grid(config = pg_current_config())
}
\arguments{
\item{config}{A \code{pg_config} object. Defaults to \code{\link[=pg_current_config]{pg_current_config()}}.}
}
\value{
SpatRaster
}
\description{
Create a raster with PRIO-GRID ids.
}
\examples{
\dontrun{
pg <- prio_blank_grid()
}
}
