% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priogrid.R
\name{prio_config}
\alias{prio_config}
\title{prio_config}
\usage{
prio_config()
}
\value{
Returns a list of named lists containing configuration
info.
}
\description{
A helper function that returns the standard configuration for
making priogrid.
}
\details{
Useful for substituting the standard functions with new ones,
or only doing certain variables.
}
