% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{prio_polygonize_grid}
\alias{prio_polygonize_grid}
\title{Polygonize grid}
\usage{
prio_polygonize_grid(rastergrid)
}
\arguments{
\item{rastergrid}{A raster}
}
\value{
A polygon grid
}
\description{
Returns a polygon version of PRIOGrid, meaning a grid of square polygons
}
