% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{prio_raster_layers}
\alias{prio_raster_layers}
\title{Raster layers}
\usage{
prio_raster_layers(raster)
}
\description{
Returns a list of n rasters where n = length(unique(values(raster)))
(Might be) useful for partitioning a raster into separate instances for
more effective computation.
}
