% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priogrid-package.R
\docType{package}
\name{priogrid-package}
\alias{priogrid}
\alias{priogrid-package}
\title{priogrid: PRIOGRID}
\description{
An R-package for collecting and standardizing open spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://prio-data.github.io/priogrid/}
  \item Report bugs at \url{https://github.com/prio-data/priogrid/issues}
}

}
\author{
\strong{Maintainer}: Jonas Vestby \email{jonves@prio.org} (\href{https://orcid.org/0000-0002-3054-977X}{ORCID})

Authors:
\itemize{
  \item Jonas Vestby \email{jonves@prio.org} (\href{https://orcid.org/0000-0002-3054-977X}{ORCID})
  \item Andreas Forø Tollefsen \email{andreas@prio.org} (\href{https://orcid.org/0000-0002-0864-5132}{ORCID})
}

Other contributors:
\itemize{
  \item Karim Bahgat \email{karim.bahgat@outlook.com} (Co-author of PG2) [contributor]
  \item Peace Research Institute Oslo (PRIO) [copyright holder]
  \item Håvard Strand \email{hs@prio.org} [contributor]
  \item Scott Gates \email{scott@prio.org} [contributor]
  \item Halvard Buhaug \email{halvard@prio.org} [contributor]
  \item Kristine Helskog \email{krihel@prio.org} [contributor]
  \item Garret Benz \email{garben@prio.org} [contributor]
  \item Ingvild Kjelstrup Smestad \email{ingviksm@ifi.uio.no} [contributor]
  \item Tiril Rahn \email{tiril.rahn@univ.ox.ac.uk} [contributor]
  \item Christian Oswald \email{christian.oswald@unibw.de} [contributor]
  \item Peder Landsverk \email{pglandsverk@gmail.com} [contributor]
  \item Vilde Bergstad Larsen \email{villar@prio.org} [contributor]
}

}
\keyword{internal}
