% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{raster_to_pg}
\alias{raster_to_pg}
\title{raster_to_pg}
\usage{
raster_to_pg(rast, aggregation_function = "mean", resampling_method = "")
}
\arguments{
\item{rast}{a raster object}

\item{aggregation_function}{see raster::aggregate fun}

\item{resampling_method}{see raster::disaggregate, default is ''}
}
\value{
A raster with the same resolution, crs, and extent as PRIO-GRID.
}
\description{
Aggregates or disaggregates a raster to conform with the PRIO-GRID resolution.
}
