% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{raster_to_pgtibble}
\alias{raster_to_pgtibble}
\title{Convert a pgraster to tabular format}
\usage{
raster_to_pgtibble(r)
}
\arguments{
\item{r}{An input raster that must have the same projection, resolution, and extent as PRIO-GRID}
}
\value{
tibble
}
\description{
Convert a pgraster to tabular format
}
\examples{
pg <- prio_blank_grid()
names(pg) <- "test"
raster_to_pgtibble(pg)
}
