% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{raster_to_table}
\alias{raster_to_table}
\title{raster_to_table}
\usage{
raster_to_table(raster)
}
\value{
A tibble with columns x, y, mydate, and value
}
\description{
Converts a raster to a tibble with x, y, mydate and raster value
}
