% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{raster_to_tibble}
\alias{raster_to_tibble}
\title{raster_to_table}
\usage{
raster_to_tibble(rast, add_pg_index = FALSE)
}
\arguments{
\item{rast}{a raster object}

\item{add_pg_index}{boolean, whether to add a pgid column or not.}
}
\value{
A tibble with columns x, y, and value.
}
\description{
Converts a raster to a tibble with x, y, mydate and raster value
}
