% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rasterextent_to_pg}
\alias{rasterextent_to_pg}
\title{rasterextent_to_pg}
\usage{
rasterextent_to_pg(rast)
}
\arguments{
\item{rast}{a raster object}
}
\value{
A raster with same extent and crs as PRIO-GRID.
}
\description{
Resamples a raster using the nearest neighbor method to make extent the same as PRIO-GRID.
}
