% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{read_naturalearth_10m_land}
\alias{read_naturalearth_10m_land}
\title{Reads the Natural Earth 10m Physical Vector Data}
\usage{
read_naturalearth_10m_land()
}
\value{
an object of class sf
}
\description{
Reads the Natural Earth 10m Physical Vector Data
}
\references{
\insertRef{naturalearthLand10mPhysical2024}{priogrid}
}
