% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{read_naturalearth_50m_land}
\alias{read_naturalearth_50m_land}
\title{read_naturalearth_10m_physical}
\usage{
read_naturalearth_50m_land()
}
\value{
an object of class sf
}
\description{
Reads the Natural Earth 10m Physical Vector Data
}
