% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ruggedterrain.R
\name{read_ruggedterrain}
\alias{read_ruggedterrain}
\title{Reads the Global Multi-resolution Terrain Elevation Data (GMTED2010)
Unzips the folder and reads the .shp file as sf object}
\usage{
read_ruggedterrain()
}
\value{
an object of class sf
}
\description{
Reads the Global Multi-resolution Terrain Elevation Data (GMTED2010)
Unzips the folder and reads the .shp file as sf object
}
\references{
\insertRef{danielsonGlobalMultiresolutionTerrain2011}{priogrid}
}
