% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{read_shdi}
\alias{read_shdi}
\title{Reads in SHDI data and fixes empty SHDI geometries}
\usage{
read_shdi(
  shdi_csv = read_shdi_csv(),
  shp = read_shdi_shapefile(),
  geoboundaries = read_geoboundaries(),
  fix_empty = TRUE
)
}
\arguments{
\item{shdi_csv}{A data frame containing SHDI data as returned by
\code{\link{read_shdi_csv}}.}

\item{shp}{An \code{sf} object containing SHDI geometries as returned by
\code{\link{read_shdi_shapefile}}.}

\item{geoboundaries}{An \code{sf} object containing administrative boundary
geometries as returned by \code{\link{read_geoboundaries}}.}

\item{fix_empty}{A boolean whether or not to fix SHDI empty geometries. Defaults to TRUE.}
}
\value{
An \code{sf} object containing SHDI data and geometries.
}
\description{
This function processes GlobalDataLab Subnational Human Development Index (SHDI)
data, resolving missing or empty geometries using geoBoundaries crosswalks, and
returns a sf data.frame
}
\examples{
\dontrun{
# Generate PRIO-GRID SHDI
shdi <- read_shdi()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{read_shdi_csv}},
\code{\link{read_shdi_shapefile}},
\code{\link{read_geoboundaries}},
\code{\link{prio_blank_grid}}
}
