% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{read_shdi_shapefile}
\alias{read_shdi_shapefile}
\title{Read GlobalDataLab SHDI Shapefile Data}
\usage{
read_shdi_shapefile()
}
\value{
An \code{sf} object containing SHDI geometries with GlobalDataLab
subnational identifiers.
}
\description{
Downloads from a PRIO hosted mirror.
}
\details{
This function reads the GlobalDataLab Subnational Human Development Index (SHDI)
shapefile dataset (version 7.0) from local storage. The shapefile is extracted
from a compressed archive and returned as an \code{sf} object with subnational
geometries linked by GlobalDataLab codes.
}
\examples{
\dontrun{
shdi_shp <- read_shdi_shapefile()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{get_pgfile}} for file retrieval,
\code{\link[sf]{st_read}} for spatial data input
}
