% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_worldpop.R
\name{read_worldpop_migration_flows}
\alias{read_worldpop_migration_flows}
\title{Reads Worldpop Migration Flows data}
\usage{
read_worldpop_migration_flows(print_metadata = FALSE)
}
\value{
list
}
\description{
Note: The shapefile for international flowlines does not seem to work.
}
\references{
\insertRef{ceausuSexDisaggregatedInternalInternational2021}{priogrid}
}
