% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{robust_transformation}
\alias{robust_transformation}
\title{Transform raster to PRIO-GRID format}
\usage{
robust_transformation(r, agg_fun, disagg_method = "near", cores = 1, ...)
}
\arguments{
\item{r}{Raster to transform}

\item{agg_fun}{Aggregation function, see terra::aggregate}

\item{disagg_method}{Disaggregation method, see terra::disagg}

\item{cores}{The number of cores to use when aggregating data}
}
\value{
SpatRaster
}
\description{
Transform raster to PRIO-GRID format
}
\examples{
r <- prio_blank_grid(ncol = 1440, nrow = 720)
r <- project(r, "ESRI:54009")
robust_transformation(r, agg_fun = "mean")
}
