% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mirca.R
\name{rollovercols}
\alias{rollovercols}
\title{Roll over Columns}
\usage{
rollovercols(data, winsize = 3)
}
\arguments{
\item{data}{A set of data}

\item{winsize}{Size of the window}

\item{exclude}{Columns to exclude}

\item{shift}{Shift the window. Ex. size 3 / shift 1 gives sum FROM col instead of around}
}
\description{
Calculates the sum of a rolling window over a range of columns.
Only considers columns not specified in "exclude"
}
\details{
Assumes that the columns are sorted
}
