% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_pg_options}
\alias{set_pg_options}
\title{Set package-options for PRIO-GRID}
\usage{
set_pg_options(opt_list)
}
\arguments{
\item{opt_list}{A named list with the option name and option value. Parameters:
ncol = number of columns in prio-grid output raster
nrow = number of rows in prio-grid output raster
ext = extent of prio-grid output raster
crs = projections of prio-grid output raster
rawfolder = folder to download and store rawdata}
}
\description{
Set package-options for PRIO-GRID
}
\examples{
#set_pg_options(list("rawfolder" = "~/priogrid_data/"))
}
