% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{specification_lookup}
\alias{specification_lookup}
\title{lookup}
\usage{
specification_lookup(type, what)
}
\description{
Returns an entry from a specification list, which is located in
the package's extdata folder. This function is used to retrieve data about
how to read/write file extensions (mm).
}
