% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_speibase.R
\name{spei.nc}
\alias{spei.nc}
\title{Compute SPEI from very large arrayed data sets in netCDF format.}
\usage{
spei.nc(
  sca,
  inPre,
  outFile,
  inEtp = NA,
  title = NA,
  comment = NA,
  version = NA,
  inMask = NA,
  block = 18,
  tlapse = NA
)
}
\arguments{
\item{sca}{Integer. The time scale of the SPEI.}

\item{inPre}{Character vector. Path to the input precipitation netCDF file.}

\item{outFile}{Character vector. Path of the output netCDF file.}

\item{inEtp}{Character vector. Path to the input evapotranspiration netCDF file.}

\item{title}{Character vector. Title of the dataset.}

\item{comment}{Character vector. A comment.}

\item{version}{Character vector. Dataset version number.}

\item{inMask}{Character vector. Path to a netCDF mask file.}

\item{block}{Integer. Number of latitude blocks to be processed at the
same time. Must be an integer dividend of 360.}

\item{tlapse}{Numeric. Temperature lapse rate. Default NA.}
}
\value{
Computes the SPEI time series and stores it in outFile following
the same data structure of inPre.
}
\description{
Compute SPEI from very large arrayed data sets in netCDF format.
}
\section{References}{

Code from Beguería S. (2017) SPEIbase: R code used in generating the SPEI global database, doi:10.5281/zenodo.834462.
}

