% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_speibase.R
\name{speibaseN}
\alias{speibaseN}
\title{Make SPEIbase data PRIO-GRID compatible}
\usage{
speibaseN(interval, time_agg_fun, config = pg_current_config())
}
\arguments{
\item{interval}{Integer. The month interval to calculate SPEI over. E.g., if the interval is
6, then monthly precipitation and potential evapotranspiration are aggregated over the last 6 months,
then the SPEI (anomaly) is calculated. This is done for each month.}

\item{time_agg_fun}{Character. Either "mean" or "max". Original data is monthly, so if PRIO-GRID is lower
resolution, then we need to aggregate over time. Currently, mean and max functions are implemented.}

\item{config}{A \code{pg_config} object. Defaults to \code{\link{pg_current_config}()}.}
}
\value{
A \code{SpatRaster} object (from the \pkg{terra} package) with spatio-temporal
resolution as defined in PRIO-GRID.
}
\description{
The function takes monthly SPEIbase data and transforms it to the temporal
resolution defined by PRIO-GRID date intervals (which may be quarterly, yearly,
or other intervals), while also performing spatial aggregation to the PRIO-GRID resolution.
}
\examples{
\dontrun{
# SPEI-6 averaged over PRIO-GRID time-interval using the mean.
spei6_mean <- speibaseN(interval = 6, time_agg_fun = "mean")
}

}
\references{
\insertRef{begueriaSPEIbase2024}{priogrid}
\insertRef{begueriaStandardizedPrecipitationEvapotranspiration2014}{priogrid}
}
\seealso{
\code{\link{read_speibase}} for reading and calculating SPEIbase,
\code{\link{read_cru_pet}} for input-data to SPEIbase,
\code{\link{read_cru_pre}} for input-data to SPEIbase,
\code{\link{pg_date_intervals}} for PRIO-GRID temporal boundaries,
\code{\link{robust_transformation}} for spatial aggregation,
\code{\link{get_pgfile}} for file management
}
