% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{stepwiseAggregate}
\alias{stepwiseAggregate}
\title{stepwiseAggregate}
\usage{
stepwiseAggregate(data, minres, fun = priogrid::quickmode,
  firstpass = TRUE)
}
\description{
This is a function for aggregating huge rasters.
The stepwise aggregation makes it possible to
aggregate huge rasters by not having to hold
them in memory in their entirety.
}
\details{
Use the function "hugeAggregate" instead of this one,
since this one will output a raster with an arbitrary
instead of a target resolution.
}
