% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{unfinished_downloads}
\alias{unfinished_downloads}
\title{Prints data.frame with unfinished downloads}
\usage{
unfinished_downloads()
}
\value{
data.frame
}
\description{
\code{\link[=download_pg_rawdata]{download_pg_rawdata()}} stores unfinished downloads if it is interrupted by
the user or the server. Use this function to see which files that failed
to completely download. This information is also used internally by \code{\link[=download_pg_rawdata]{download_pg_rawdata()}}
to resume file downloads.
}
\examples{
unfinished_downloads()
}
