% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_pg_config}
\alias{validate_pg_config}
\title{Validate a pg_config object}
\usage{
validate_pg_config(cfg)
}
\arguments{
\item{cfg}{A list with config fields.}
}
\value{
NULL (invisibly). Stops with an error on invalid input.
}
\description{
Validate a pg_config object
}
\keyword{internal}
