% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_priogrid.R
\name{writeColored}
\alias{writeColored}
\title{writeColored}
\usage{
writeColored(x)
}
\arguments{
\item{x}{A string}
}
\description{
writeLines wrapper that colors certain things in the output
}
\examples{
writeColored("ERROR: this will be red!")

}
