% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{yearly_brick}
\alias{yearly_brick}
\title{Create RasterBrick consisting of one RasterLayer for each year}
\usage{
yearly_brick(data, variable, raster.fun)
}
\arguments{
\item{data}{sf object.}

\item{variable}{Variable to rasterize, character string.}

\item{raster.fun}{Function to rasterize by, character string.}
}
\description{
Create RasterBrick consisting of one RasterLayer for each year
}
