% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{zip_file}
\alias{zip_file}
\title{Unzip a File to a Structured Directory (with Safety Check)}
\usage{
zip_file(f)
}
\arguments{
\item{f}{Character. Full path to the \code{.zip} file to be extracted.}
}
\value{
A character string indicating the full path to the directory
where the contents were (or would have been) unzipped.
}
\description{
Extracts the contents of a \code{.zip} file into a uniquely named subdirectory
in the same location as the original file. The output folder is named after
the zip file with a \verb{_unzipped} suffix.
}
\details{
If the target directory already exists and contains files or subdirectories,
the function skips unzipping and returns the existing directory path.
}
\examples{
\dontrun{
zip_path <- "/path/to/hildap_vGLOB-1.0_geotiff.zip"
unzipped_dir <- zip_file(zip_path)
list.files(unzipped_dir)
}

}
