% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{bdist2}
\alias{bdist2}
\title{Calculates bdist2 variable for a single time slice.}
\usage{
bdist2(measurement_date, cshp = read_cshapes(), past_result = NULL)
}
\arguments{
\item{measurement_date}{A single date as Date object}

\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}

\item{past_result}{A past result of bdist2.}
}
\value{
list with SpatRast with bdist2, and Sf objects with country boundaries and shared_borders.
}
\description{
bdist2 gives the spherical distance (in kilometer) from the cell centroid to
the border of the nearest neighboring country, regardless of whether the nearest
country is located across international waters. Based on cShapes dataset.`
}
\details{
The variable can be generated with up to daily temporal resolution. It does
take quite a long time to estimate, but can use a past result if there are
no changes in borders.
}
\examples{
# bdist2 <- bdist2(as.Date("2010-01-01"))
}
\references{
\insertRef{schvitzMappingInternationalSystem2022}{priogrid}
}
