% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{calc_pg}
\alias{calc_pg}
\title{Calculates all PRIO-GRID variables}
\usage{
calc_pg(varnames = NULL, overwrite = FALSE)
}
\arguments{
\item{varnames}{Vector with names of variables in PRIOGRID. See \link{pgvariables}. If NULL, then all variables are used.}

\item{overwrite}{Boolean, if false, will ignore updating variables that already exist in the result folder}
}
\description{
In PRIO-GRID, all functions that are named "gen_something()" are functions that
return final variables as a SpatRaster (or a stack of these). This function
calculates all variables and store them in "path/to/your/rawfolder/priogrid/version/{options_hash}/variable_name.rds".
}
\examples{
\dontrun{
  calc_pg("ne_disputed_area_share")
  r <- load_pgvariable()
}

}
