% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_traveltime.R
\name{calc_traveltime}
\alias{calc_traveltime}
\title{Generate travel time to nearest major city}
\usage{
calc_traveltime(aggregation_function)
}
\arguments{
\item{aggregation_function}{function used to aggregate values. Either an actual function,
or for the following, their name: "mean", "max", "min", "median", "sum", "modal", "any",
"all", "prod", "which.min", "which.max", "table", "sd" (sample standard deviation) and
"std" (population standard deviation)}
}
\description{
Aggregates the high resolution travel time data to PRIO-GRID level
}
\examples{
# r <- calc_traveltime(aggregation_function = "median")


}
\references{
\insertRef{nelsonTravelTimeMajor2008}{priogrid}
}
