% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_pg_indices}
\alias{create_pg_indices}
\title{Create matrix with index numbering conventions as for PRIO-GRID.}
\usage{
create_pg_indices(ncol = pgoptions$get_ncol(), nrow = pgoptions$get_nrow())
}
\arguments{
\item{ncol}{The number of columns in the grid.}

\item{nrow}{The number of rows in the grid.}
}
\value{
A ncol*nrow matrix with integer indices.
}
\description{
Create matrix with index numbering conventions as for PRIO-GRID.
}
\examples{
pg <- create_pg_indices()
}
