% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_pg_rawdata}
\alias{download_pg_rawdata}
\title{Download the raw-data for PRIO-GRID}
\usage{
download_pg_rawdata(overwrite = FALSE, file_info = NULL)
}
\arguments{
\item{overwrite}{Whether or not to download and overwrite files already in local folder.}

\item{file_info}{A data.frame with the same structure as the result from \code{\link[=pg_rawfiles]{pg_rawfiles()}}. If file_info is null (default),
then file_info will be all data returned from \code{\link[=pg_rawfiles]{pg_rawfiles()}}.}
}
\description{
Before running this, you need to set the folder using pgoptions$set_rawfolder("path/to/folder")
}
