% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_pg_rawdata}
\alias{download_pg_rawdata}
\title{Download the raw-data for PRIO-GRID}
\usage{
download_pg_rawdata(
  file_info = NULL,
  overwrite = FALSE,
  resume = TRUE,
  max_retry = 10
)
}
\arguments{
\item{file_info}{A data.frame with the same structure as the result from \code{\link[=pg_rawfiles]{pg_rawfiles()}}. If file_info is null (default),
then file_info will be all data returned from \code{\link[=pg_rawfiles]{pg_rawfiles()}}.}

\item{overwrite}{Whether or not to download and overwrite files already in local folder.}

\item{resume}{If true, will also download files that did not finish download last time the function was run.}
}
\value{
data.frame Download summary
}
\description{
Before running this, you need to set the folder using pgoptions$set_rawfolder("path/to/folder")
}
\examples{
files_to_download <- pg_rawfiles() |> dplyr::filter(id == "ec3eea2e-6bec-40d5-a09c-e9c6ff2f8b6b")
# download_pg_rawdata(overwrite = TRUE, file_info = files_to_download)
}
