% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{extract_bib_elements}
\alias{extract_bib_elements}
\title{Helper function to parse bibliography elements based on a source}
\usage{
extract_bib_elements(citation_liststr, bib_element = "author", ...)
}
\arguments{
\item{citation_liststr}{A semi-colon separated list of bibliography keys from \link{pgsources}}

\item{bib_element}{Supports author, journal, year, and title.}

\item{...}{}
}
\value{
list with BibEntry-element classes or vectors with character strings
}
\description{
In \link{pgsources}, citations are semi-colon separated. This function splits
these into individual keys and get the bibliography element from each key.
}
\examples{
extract_bib_elements(pgsources$citation_keys[1])
}
