% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_bdist2}
\alias{gen_bdist2}
\title{gen_bdist2}
\usage{
gen_bdist2(input_folder, ...)
}
\arguments{
\item{interval}{see by argument in base::seq.Date. default is 1 year.}

\item{enddate}{a date. option to add more years than the last entry in the input data}

\item{startdate}{a date. filters the input data to start at the given date.}
}
\description{
Takes bdist2_changes and translates it into a regular time-series.
}
