% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{gen_bdist2}
\alias{gen_bdist2}
\title{Generate bdist2 variable}
\usage{
gen_bdist2(cshp = read_cshapes())
}
\arguments{
\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\value{
SpatRast
}
\description{
bdist2 gives the spherical distance (in kilometer) from the cell centroid to
the border of the nearest neighboring country, regardless of whether the nearest
country is located across international waters. Based on cShapes dataset.`
}
\details{
The variable can be generated with up to daily temporal resolution. It does
take quite a long time to estimate, but uses a past result if there are
no changes in borders.
}
\examples{
# bdist2 <- gen_bdist2()
}
\references{
\insertRef{schvitzMappingInternationalSystem2022}{priogrid}
}
