% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_changed_areas}
\alias{gen_changed_areas}
\title{gen_changed_areas}
\usage{
gen_changed_areas(fname, numCores = 1, quiet = TRUE)
}
\arguments{
\item{fname}{File path to Weidmann cshapes data}

\item{numCores}{Number of cores to use in calculation. Windows-users can only use 1. Using parallel::mclapply.}

\item{quiet}{Whether or not sf::st_ functions should print warnings.}
}
\description{
Takes the weidmann cshapes data set and returns a
list of sf dataframes that corresponds to the areas where borders changed from one month to the next.
}
