% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cmr.R
\name{gen_cmr_sd}
\alias{gen_cmr_sd}
\title{cmr_sd}
\usage{
gen_cmr_sd(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Measures the prevalence of child malnutrition within each cell. The original pixel value
is the percent of children under the age of 5 that are malnourished. This variable gives the standard deviation
of original pixel values within each cell. Based on raster data from the SEDAC Global Poverty Mapping
project. The indicator is a snapshot for the year 2000 only.

Link to original data: https://sedac.ciesin.columbia.edu/data/set/povmap-global-subnational-prevalence-child-malnutrition

Please cite: Center for International Earth Science Information Network - CIESIN - Columbia University. 2005. Poverty Mapping Project: Global Subnational Prevalence of Child Malnutrition. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). https://doi.org/10.7927/H4K64G12.
}
