% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cru.R, R/data_crupre.R
\name{gen_cru_pre}
\alias{gen_cru_pre}
\title{Generate PRIO-GRID Compatible CRU Precipitation Data}
\usage{
gen_cru_pre()

gen_cru_pre()
}
\value{
A \code{SpatRaster} object with PRIO-GRID spatio-temporal resolution.

A \code{SpatRaster} object with PRIO-GRID spatio-temporal resolution.
}
\description{
This function processes CRU TS v4.09 precipitation (PRE)
data by aggregating it temporally and spatially to match PRIO-GRID
specifications.

This function processes CRU TS v4.09 precipitation (PRE)
data by aggregating it temporally and spatially to match PRIO-GRID
specifications.
}
\examples{
\dontrun{
r <- gen_cru_pre()
}

\dontrun{
r <- gen_cru_pre()
}

}
\seealso{
\code{\link{read_cru_pre}}, \code{\link{pg_date_intervals}},
\code{\link{robust_transformation}}, \code{\link{get_pgfile}}

\code{\link{read_cru_pre}}, \code{\link{pg_date_intervals}},
\code{\link{robust_transformation}}, \code{\link{get_pgfile}}
}
