% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{gen_cshapes_cover_share}
\alias{gen_cshapes_cover_share}
\title{Generate cshapes_cover_share variable}
\usage{
gen_cshapes_cover_share(cshp = read_cshapes())
}
\arguments{
\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\value{
SpatRast
}
\description{
The variable can be generated with up to daily temporal resolution.
}
\examples{
# cshapes_cover_share <- gen_cshapes_cover_share()
}
