% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{gen_cshapes_gwcode}
\alias{gen_cshapes_gwcode}
\title{Generate cshapes_gwcode variable}
\usage{
gen_cshapes_gwcode(cshp = read_cshapes())
}
\arguments{
\item{cshp}{The CShapes dataset, for instance as given by \code{\link[=read_cshapes]{read_cshapes()}}}
}
\value{
SpatRast
}
\description{
The variable can be generated with up to daily temporal resolution.
}
\examples{
# gwcode <- gen_cshapes_gwcode()
}
\references{
\insertRef{schvitzMappingInternationalSystem2022}{priogrid}
}
