% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_diamprim_y}
\alias{gen_diamprim_y}
\title{Generate yearly primary diamond presence dummy.}
\usage{
gen_diamprim_y(diamond_data)
}
\arguments{
\item{diamond_data}{DIADATA shapefile from the PRIO Diamond Resources dataset.}
}
\description{
Generate yearly primary diamond presence dummy.
}
