% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_diamprim_y}
\alias{gen_diamprim_y}
\title{diamprim_y}
\usage{
gen_diamprim_y(input_folder, variable = "diamprim_y")
}
\arguments{
\item{input_folder}{path to [pg-folder].}

\item{variable}{one of c("diamprim_y", "diamsec_y")}
}
\description{
Generate dummy variable for whether
primary (kimberlite or lamproite) diamond deposits have
been found within the given grid cell for any given year,
based on the Diamond Resources dataset v1a.
This variable only codes deposits that have a known year of
discovery or start of production. For a complete picture,
these data must therefore be combined with the diamprim_s data.
Available only until 2005.

Link to original data: https://www.prio.org/Data/Geographical-and-Resource-Datasets/Diamond-Resources/

Please cite: Gilmore, Elisabeth; Nils Petter Gleditsch, Päivi Lujala & Jan Ketil Rød, 2005. ‘Conflict Diamonds: A New Dataset’, Conflict Management and Peace Science 22(3): 257–292
Lujala, Päivi; Nils Petter Gleditsch & Elisabeth Gilmore, 2005. ‘A Diamond Curse? Civil War and a Lootable Resource’, Journal of Conflict Resolution 49(4): 538–562.
}
