% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_diamsec_y}
\alias{gen_diamsec_y}
\title{Generate yearly secondary diamond presence dummy.}
\usage{
gen_diamsec_y(diamond_data)
}
\arguments{
\item{diamond_data}{DIADATA shapefile from the PRIO Diamond Resources dataset.}
}
\description{
Generate yearly secondary diamond presence dummy.
}
