% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_drug_y}
\alias{gen_drug_y}
\title{drug_y}
\usage{
gen_drug_y(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether large-scale
drug cultivation (coca bush, opium poppy, or cannabis) is ongoing
within the given grid cell, based on the DRUGDATA dataset.
Available only until 2002.

Link to original data: http://www.paivilujala.com/drugdata.html

Please cite: Buhaug, Halvard & Päivi Lujala 2005. Accounting for Scale: Measuring Geography in Quantitative Studies of Civil War. Political Geography 24: 399-418.
}
