% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_epr_regional}
\alias{gen_epr_regional}
\title{epr_regional}
\usage{
gen_epr_regional(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Counts the number of regionally based ethnic groups
that are settled in a given grid cell in a given year; derived from
the GeoEPR 2019 data.

Link to original data: https://icr.ethz.ch/data/epr/

Please cite: Vogt, Manuel, Nils-Christian Bormann, Seraina Rüegger, Lars-Erik Cederman, Philipp Hunziker, and Luc Girardin. 2015. “Integrating Data on Ethnicity, Geography, and Conflict: The Ethnic Power Relations Data Set Family.” Journal of Conflict Resolution 59(7): 1327–42.
}
