% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{gen_esch}
\alias{gen_esch}
\title{Generate PRIO-GRID Expected Years of Schooling}
\usage{
gen_esch()
}
\value{
A \code{SpatRaster} object containing PRIO-GRID expected years of schooling.
}
\description{
Convenience wrapper around \code{\link{shdi}} that generates a PRIO-GRID–aligned
raster of expected years of schooling.
}
\examples{
\dontrun{
esch_pg <- gen_esch()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{shdi}}
}
