% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_excluded}
\alias{gen_excluded}
\title{excluded}
\usage{
gen_excluded(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate variable that counts the number of excluded and politically relevant
ethnic groups that are settled in a given grid cell in a given year.
}
