% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nordhaus.R
\name{gen_gcp_mer}
\alias{gen_gcp_mer}
\title{Generate gcp_mer variable}
\usage{
gen_gcp_mer(input_folder, variable = "GCP_MER")
}
\arguments{
\item{gecon_data}{G-Econ v. 4.0 excel data}
}
\description{
Generate gross cell product (USD) variable, based on the G-Econ dataset v. 4.0.
}
