% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nordhaus.R
\name{gen_gcp_qual}
\alias{gen_gcp_qual}
\title{Generate gcp_qual variable}
\usage{
gen_gcp_qual(gecon_data)
}
\arguments{
\item{gecon_data}{G-Econ v. 4.0 excel data}
}
\description{
Generate quality of GCP values variable,
based on the G-Econ dataset v. 4.0.
}
