% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_gem_y}
\alias{gen_gem_y}
\title{Generate yearly gem dummy.
Function to generate yearly gem presence dummy for records with known discovery or start of production year.}
\usage{
gen_gem_y(gem_data)
}
\arguments{
\item{gemdata}{GEMDATA shapefile.}
}
\description{
Generate yearly gem dummy.
Function to generate yearly gem presence dummy for records with known discovery or start of production year.
}
