% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geoEPR.R
\name{gen_geoepr_reg_excluded}
\alias{gen_geoepr_reg_excluded}
\title{PRIO-GRID cells with any regionally based politically excluded groups}
\usage{
gen_geoepr_reg_excluded(
  excluded = c("DISCRIMINATED", "POWERLESS", "SELF-EXCLUSION")
)
}
\arguments{
\item{excluded}{A character vector with the EPR status codes to include in the "excluded" group.
The status codes in EPR is "MONOPOLY", "DISCRIMINATED", "POWERLESS", "DOMINANT", "SENIOR PARTNER",
"JUNIOR PARTNER", "IRRELEVANT", "SELF-EXCLUSION", and "STATE COLLAPSE"}
}
\value{
SpatRast
}
\description{
PRIO-GRID cells with any regionally based politically excluded groups
}
\examples{
# r <- gen_geoepr_reg_excluded()

}
\references{
\insertRef{wucherpfennigPoliticallyRelevantEthnic2011}{priogrid}
\insertRef{vogtIntegratingDataEthnicity2015}{priogrid}
}
