% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geopko.R
\name{gen_geopko_operations_count}
\alias{gen_geopko_operations_count}
\title{Generate Geo-PKO count}
\usage{
gen_geopko_operations_count()
}
\value{
SpatRast
}
\description{
Takes the count of peacekeeping operations within each PRIO-GRID cell per year
}
\examples{
# r <- gen_geopko_operations_count()

}
\references{
\insertRef{cilMappingBlueHelmets2020}{priogrid}
}
