% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geopko.R
\name{gen_geopko_troops_count}
\alias{gen_geopko_troops_count}
\title{Generate the number of troops from Geo-PKO}
\usage{
gen_geopko_troops_count()
}
\value{
SpatRast
}
\description{
Takes the number of peace keeping troops within each PRIO-GRID cell
}
\examples{
# t <- gen_pko_troops_count
}
\references{
\insertRef{cilMappingBlueHelmets2020}{priogrid}
}
