% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ghsl.R
\name{gen_ghsl_population_grid}
\alias{gen_ghsl_population_grid}
\title{Generate GHSL GHS Population Grid}
\usage{
gen_ghsl_population_grid()
}
\description{
This aggregates the high-resolution population grid to PRIO-GRID level across all 5-year intervals (1975 - 2030).
}
\details{
This does take some time.

A slight nearest neighbor resampling was applied to get the exact PRIO-GRID extent.
}
\examples{
# r <- gen_ghsl_population_grid()
}
\references{
\insertRef{schiavinaGHSPOPR2023AGHS2023}{priogrid}
}
