% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{gen_gnic}
\alias{gen_gnic}
\title{Generate PRIO-GRID Gross National Income per Capita}
\usage{
gen_gnic()
}
\value{
A \code{SpatRaster} object containing PRIO-GRID GNI per capita values.
}
\description{
Convenience wrapper around \code{\link{shdi}} that generates a PRIO-GRID–aligned
raster of gross national income per capita.
}
\examples{
\dontrun{
gnic_pg <- gen_gnic()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{shdi}}
}
