% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldplacer_y}
\alias{gen_goldplacer_y}
\title{goldplacer_y}
\usage{
gen_goldplacer_y(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
placer gold deposits have been found within the given
grid cell, based om the dGOLD_L subset of the
GOLDATA dataset v1.2. Available only until 2012.
}
