% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldplacer_y}
\alias{gen_goldplacer_y}
\title{Generate yearly dummy identifying lootable placer gold deposits.}
\usage{
gen_goldplacer_y(input_folder)
}
\description{
Generate yearly dummy identifying lootable placer gold deposits.
}
