% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldvein_s}
\alias{gen_goldvein_s}
\title{goldvein_s}
\usage{
gen_goldvein_s(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Generate dummy variable for whether
vein gold deposits have been found within the given grid cell,
based on the dGOLD_NL subset of the GOLDATA v1.2 dataset.
This variable only codes those deposits that do not have a known
year of discovery or start of production. For a complete picture,
these data must therefore be combined with the goldvein_y data.

Link to original data: https://www.researchgate.net/publication/281849073_GOLDATA_12_v

Please cite: Balestri, Sara, 2015. GOLDATA: The Gold deposits dataset codebook, Version 1.2. UCSC-Cognitive Science and Communication Research Centre WP 02/15, Milan. doi:10.13140/RG.2.1.1730.8648
Balestri, Sara, 2012. Gold and civil conflict intensity: evidence from a spatially disaggregated analysis, Peace Economics. Peace Science and Public Policy, 18(3): 1-17. doi:10.1515/peps-2012-0012.
}
