% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resource.R
\name{gen_goldvein_y}
\alias{gen_goldvein_y}
\title{Generate yearly dummy identifying non-lootable vein gold deposits.}
\usage{
gen_goldvein_y(golddata_nl)
}
\arguments{
\item{golddata_nl}{GOLDATA 1.2 v non-lootable gold shapefile.}
}
\description{
Generate yearly dummy identifying non-lootable vein gold deposits.
}
