% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_gwcode}
\alias{gen_gwcode}
\title{gen_gwcode}
\usage{
gen_gwcode(fname, numCores = 1, quiet = TRUE)
}
\arguments{
\item{fname}{File path to Weidmann cshapes data}

\item{numCores}{Number of cores to use in calculation. Windows-users can only use 1. Using parallel::mclapply.}

\item{quiet}{Whether or not sf::st_ functions should print warnings.}
}
\description{
Takes the weidmann cshapes data set and returns a
rasterstack for each year that codes the country ownership of each cell
as it looked like 1 January each year.
}
