% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_linight.R
\name{gen_linight_mean}
\alias{gen_linight_mean}
\title{Generate Li Nighttime Light}
\usage{
gen_linight_mean()
}
\value{
SpatRast
}
\description{
This aggregates the high-resolution Li Nighttime Light to PRIO-GRID level across all years (1992-2021).
}
\details{
A slight nearest neighbor resampling was applied to get the exact PRIO-GRID extent.This is done in \code{\link[=read_linight]{read_linight()}}.
}
\examples{
# r <- gen_linight_mean()
}
\references{
\insertRef{liHarmonizedGlobalNighttime2020}{priogrid}
}
