% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{gen_msch}
\alias{gen_msch}
\title{Generate PRIO-GRID Mean Years of Schooling}
\usage{
gen_msch()
}
\value{
A \code{SpatRaster} object containing PRIO-GRID mean years of schooling.
}
\description{
Convenience wrapper around \code{\link{shdi}} that generates a PRIO-GRID–aligned
raster of mean years of schooling.
}
\examples{
\dontrun{
msch_pg <- gen_msch()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{shdi}}
}
