% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shdi.R
\name{gen_msch}
\alias{gen_msch}
\title{msch}
\usage{
gen_msch(input_folder, add_missing_geometries = TRUE)
}
\arguments{
\item{input_folder}{path to [pg-folder].}

\item{add_missing_geometries}{if `TRUE`, adds geometries that are missing in the original data. Replacement geometries are collected from cshapes or GADM (https://gadm.org/maps.html).}
}
\description{
Gives the mean years of schooling for population aged 25+ within each grid cell, based on SHDI data v. 4. Available from 1990 to 2018.
}
\details{
Links to original data: https://globaldatalab.org/shdi/shapefiles/ and https://globaldatalab.org/shdi/download_files/.

Please cite: Smits, J., Permanyer, I. The Subnational Human Development Database. Sci Data 6, 190038 (2019). https://doi.org/10.1038/sdata.2019.38
}
