% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{gen_naturalearth_cover}
\alias{gen_naturalearth_cover}
\title{Whether or not a grid-cell intersects with land (NaturalEarth)}
\usage{
gen_naturalearth_cover()
}
\description{
Takes the NaturalEarth 50m raster and returns
a raster-mask that is true for the grid cells that intersects with land.
}
