% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{gen_naturalearth_cover_share}
\alias{gen_naturalearth_cover_share}
\title{Share of grid-cell that intersects with land (NaturalEarth)}
\usage{
gen_naturalearth_cover_share()
}
\description{
Takes the NaturalEarth 10m raster and returns a raster
in PRIO-GRID resolution with the share of the cell that intersects with land.
}
