% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{gen_ne_disputed_area_share}
\alias{gen_ne_disputed_area_share}
\title{Generate raster for disputed areas by type of dispute}
\usage{
gen_ne_disputed_area_share()
}
\value{
terra SpatRaster
}
\description{
Creates a raster in PRIO-GRID resolution for disputed areas from the
Natural Earth Disputed Areas dataset using all types of disputes. The timing
of the disputes are not dated.
}
\examples{
\dontrun{
  rast <- gen_ne_disputed_area_share()
}
}
\references{
\insertRef{naturalearthAdmin0Breakaway2024}{priogrid}
}
