% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshp.R
\name{gen_pgland}
\alias{gen_pgland}
\title{gen_pgland}
\usage{
gen_pgland(fname, quiet = TRUE)
}
\arguments{
\item{fname}{File path to Weidmann cshapes data}

\item{quiet}{Whether or not sf::st_ functions should print warnings.}
}
\description{
Takes the weidmann cshapes data set and returns a raster for the
grid cells that intersects with land.
}
