% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gpw.R
\name{gen_pop_gpw_max}
\alias{gen_pop_gpw_max}
\title{pop_gpw_max}
\usage{
gen_pop_gpw_max(
  input_folder,
  interpolate_time = FALSE,
  interpolate_missing = TRUE
)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}

\item{interpolate_time}{if `TRUE`, data is interpolated to all years from 2000 to 2020. See interpolate_pg_timeseries() for details.}

\item{interpolate_missing}{if `TRUE`, interpolates data to grid cells with missing values around coastline.}
}
\description{
Generate maximum population count
for years 2000, 2005, 2010, 2015 and 2020 from
the Gridded Population of the World v.4 data.

Link to original data: https://sedac.ciesin.columbia.edu/data/collection/gpw-v4.

Please cite: Center for International Earth Science Information Network - CIESIN - Columbia University. 2018. Gridded Population of the World, Version 4 (GPWv4): Population Count, Revision 11. Palisades, NY: NASA Socioeconomic Data and Applications Center (SEDAC). https://doi.org/10.7927/H4JW8BX5.
}
