% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epr.R
\name{gen_promoted}
\alias{gen_promoted}
\title{epr_promoted}
\usage{
gen_promoted(input_folder)
}
\arguments{
\item{input_folder}{Path to [pg-folder].}
}
\description{
Counts the number of politically relevant ethnic groups in the grid cell
that have increased their political status from the previous year. Group status derived from
the EPR Core 2019 dataset, and location derived from the GeoEPR 2019 data.
Status defined as group rules alone; shares power; or is excluded.

Link to original data: https://icr.ethz.ch/data/epr/

Please cite: Vogt, Manuel, Nils-Christian Bormann, Seraina Rüegger, Lars-Erik Cederman, Philipp Hunziker, and Luc Girardin. 2015. “Integrating Data on Ethnicity, Geography, and Conflict: The Ethnic Power Relations Data Set Family.” Journal of Conflict Resolution 59(7): 1327–42.
}
