% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ruggedterrain.R
\name{gen_ruggedterrain_elevation_mean}
\alias{gen_ruggedterrain_elevation_mean}
\title{Generate the average (mean) elevation}
\usage{
gen_ruggedterrain_elevation_mean()
}
\value{
SpatRast
}
\description{
Generate the average (mean) elevation
}
\examples{
# r <- gen_ruggedterrain_elevation_mean()


}
\references{
\insertRef{danielsonGlobalMultiresolutionTerrain2011}{priogrid}
}
