% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SHDI.R
\name{gen_shdi}
\alias{gen_shdi}
\title{Generate PRIO-GRID SHDI}
\usage{
gen_shdi()
}
\value{
A \code{SpatRaster} object containing PRIO-GRID SHDI values.
}
\description{
Convenience wrapper around \code{\link{shdi}} that generates a PRIO-GRID–aligned
raster of the Subnational Human Development Index (SHDI).
}
\examples{
\dontrun{
shdi_pg <- gen_shdi()
}

}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
\seealso{
\code{\link{shdi}}
}
