% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shdi.R
\name{gen_shdi}
\alias{gen_shdi}
\title{Generate Expected years schooling children aged 6 variable from SHDI}
\usage{
gen_shdi(input_folder, variable = "shdi")
}
\arguments{
\item{variable}{one of c("msch", "esch", "lifexp", "lgnic", "shdi")}

\item{path}{path to SHDI data.}
}
\description{
Average Expected years schooling children aged 6 within each grid cell.
}
