% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shdi.R
\name{gen_shdi}
\alias{gen_shdi}
\title{Generate Expected years schooling children aged 6 variable from SHDI}
\usage{
gen_shdi(input_folder, variable = "shdi", add_missing_geometries = TRUE)
}
\arguments{
\item{input_folder}{path to SHDI data.}

\item{variable}{one of c("msch", "esch", "lifexp", "lgnic", "shdi")}
}
\description{
Average Expected years schooling children aged 6 within each grid cell.
}
