% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_side.R
\name{gen_side_irrelevant}
\alias{gen_side_irrelevant}
\title{Generate SIDE Irrelevant Share}
\usage{
gen_side_irrelevant()
}
\value{
A \code{SpatRaster} object
}
\description{
Generate SIDE Irrelevant Share
}
