% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_speibase.R
\name{gen_speibase6_mean}
\alias{gen_speibase6_mean}
\title{Generate SPEI-6 with mean-temporal aggregation compatible with PRIO-GRID}
\usage{
gen_speibase6_mean()
}
\value{
A \code{SpatRaster} object (from the \pkg{terra} package) with spatio-temporal
resolution as defined in PRIO-GRID.
}
\description{
The function takes monthly SPEIbase data and transforms it to the temporal
resolution defined by PRIO-GRID date intervals (which may be quarterly, yearly,
or other intervals), while also performing spatial aggregation to the PRIO-GRID resolution.
}
\examples{
\dontrun{
# SPEI-6 averaged over PRIO-GRID time-interval using the mean.
spei6_mean <- gen_speibase6_mean()
}

}
\references{
\insertRef{begueriaSPEIbase2024}{priogrid}
\insertRef{begueriaStandardizedPrecipitationEvapotranspiration2014}{priogrid}
}
\seealso{
\code{\link{read_speibase}} for reading and calculating SPEIbase,
\code{\link{read_cru_pet}} for input-data to SPEIbase,
\code{\link{read_cru_pre}} for input-data to SPEIbase,
\code{\link{pg_date_intervals}} for PRIO-GRID temporal boundaries,
\code{\link{robust_transformation}} for spatial aggregation,
\code{\link{get_pgfile}} for file management
}
