% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_traveltime.R
\name{gen_traveltime_mean}
\alias{gen_traveltime_mean}
\title{Generate the average (mean) travel time to nearest major city}
\usage{
gen_traveltime_mean()
}
\value{
SpatRast
}
\description{
Generate the average (mean) travel time to nearest major city
}
\examples{
# r <- gen_traveltime_mean()


}
\references{
\insertRef{nelsonTravelTimeMajor2008}{priogrid}
}
