% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_traveltime.R
\name{gen_traveltime_min}
\alias{gen_traveltime_min}
\title{Generate the minimum travel time to nearest major city}
\usage{
gen_traveltime_min()
}
\value{
SpatRast
}
\description{
Generate the minimum travel time to nearest major city
}
\examples{
# r <- gen_traveltime_min()


}
\references{
\insertRef{nelsonTravelTimeMajor2008}{priogrid}
}
