% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_pgfile}
\alias{get_pgfile}
\title{Get file-path on local system to a data source in PRIO-GRID}
\usage{
get_pgfile(
  source_name,
  source_version,
  id,
  verify_checksums = pg_current_config()$verify_checksums
)
}
\arguments{
\item{source_name}{Character. The source name.}

\item{source_version}{Character. The version number.}

\item{id}{Character. The source id (UUID).}

\item{verify_checksums}{Logical. If TRUE, verifies file checksums against
stored MD5 values. Defaults to value from \code{\link[=pg_current_config]{pg_current_config()}}.}
}
\value{
file path, string
}
\description{
To look up src_name and version in PRIO-GRID, see \code{\link[=pg_rawfiles]{pg_rawfiles()}}.
}
\examples{
get_pgfile(source_name = "ETH ICR cShapes", source_version = "2.0", id = "ec3eea2e-6bec-40d5-a09c-e9c6ff2f8b6b")
}
