% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{get_spatial_hash}
\alias{get_spatial_hash}
\title{Get spatial configuration hash}
\usage{
get_spatial_hash()
}
\value{
Character string with 6-character hash
}
\description{
Creates a 6-character MD5 hash of the current spatial options from pgoptions.
}
\keyword{internal}
