% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{load_pgvariable}
\alias{load_pgvariable}
\title{Load a PRIO-GRID variable}
\usage{
load_pgvariable(varname)
}
\arguments{
\item{varname}{Character string with the variable name}
}
\value{
A SpatRaster object
}
\description{
Load a PRIO-GRID variable
}
