% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{ne_disputed_area_share}
\alias{ne_disputed_area_share}
\title{Generate raster for disputed areas by type of dispute}
\usage{
ne_disputed_area_share(type)
}
\arguments{
\item{type}{Character string indicating the disputed area type (e.g., "Disputed").}

\item{disputed_areas}{sf object of disputed areas, default is from \code{read_disputed_areas()}.}
}
\value{
terra SpatRaster
}
\description{
Creates a raster in PRIO-GRID resolution for a specified type of disputed area.
The options include: "all", "breakaway", "disputed", "geo subunit","geo unit", "indeterminate", "lease" and "overlay"
}
\examples{
\dontrun{
  # Generate raster for "Disputed" type
  disputed_raster <- ne_disputed_area_share("Disputed")
  terra::plot(disputed_raster, main = "Disputed Areas: Disputed")
}

}
\references{
\insertRef{naturalearthAdmin0Breakaway2024}{priogrid}
}
