% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pg_date_intervals}
\alias{pg_date_intervals}
\title{Get a sequence of date intervals}
\usage{
pg_date_intervals(
  start_date = pgoptions$get_start_date(),
  end_date = pgoptions$get_end_date(),
  temporal_resolution = pgoptions$get_temporal_resolution()
)
}
\arguments{
\item{from}{starting date}

\item{to}{end date}

\item{by}{increment of sequence. See details in \link[base:seq.Date]{base::seq.Date}.}
}
\value{
Date interval vector
}
\description{
The start date is the last observed date in the first interval of the intervals you
want to capture. E.g., if start date is 31 January 2010 with 1 month resolution, then the
first interval goes from 1 January 2010 to 31 January 2010. The intervals are not necessary
equal length.
}
\examples{
pg_date_intervals()
}
