% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pg_dates}
\alias{pg_dates}
\title{Get a sequence of dates}
\usage{
pg_dates(
  start_date = pgoptions$start_date(),
  end_date = pgoptions$end_date(),
  temporal_resolution = pgoptions$get_temporal_resolution()
)
}
\arguments{
\item{from}{starting date}

\item{to}{end date}

\item{by}{increment of sequence. See details in \link[base:seq.Date]{base::seq.Date}.}
}
\description{
This is a wrapper of \link[base:seq.Date]{base::seq.Date}, only with defaults that can be set in the options.
}
