% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{pg_rawfiles}
\alias{pg_rawfiles}
\title{Extract url- and file-info from PRIO-GRID metadata}
\usage{
pg_rawfiles(use_mirror = TRUE)
}
\arguments{
\item{use_mirror}{Boolean. Whether or not to use PRIO-GRID mirror.}
}
\value{
data.frame
}
\description{
Extract url- and file-info from PRIO-GRID metadata
}
