% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{pgneighbors_nnb}
\alias{pgneighbors_nnb}
\title{Get first order neighbors in PRIO-GRID, vectorized version.}
\usage{
pgneighbors_nnb(gids, ncol, nrow)
}
\arguments{
\item{gids}{The GIDs to the cells that you would like to find the neighbors of.}

\item{ncol}{The number of columns in the grid.}

\item{nrow}{The number of rows in the grid.}
}
\value{
A dataframe with a list of neighbors for each gid in gids.
}
\description{
Get first order neighbors in PRIO-GRID, vectorized version.
}
