% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{pgout_path}
\alias{pgout_path}
\title{Path to store PRIOGRID results}
\usage{
pgout_path()
}
\value{
File path
}
\description{
The path varies depending on your \link{pgoptions}:
rawfolder/priogrid/(package version)/(spatial options hash)/(temporal options hash)
}
\examples{
\dontrun{
  pgout_path()
  list.files(pgout_path())
}
}
