% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{pgsearch}
\alias{pgsearch}
\title{Search PRIO-GRID meta-data}
\usage{
pgsearch(search_string, bib_element = NULL)
}
\arguments{
\item{search_string}{}
}
\description{
Use regex to search the meta-data for the data you are interested in.
}
