% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgvariables}
\alias{pgvariables}
\title{PRIO-GRID Variables}
\format{
\subsection{\code{pgvariables}}{

A data frame with 4 rows and 5 columns:
\describe{
\item{name}{Full name of the variable.}
\item{static}{Boolean. Whether the variable varies over time or not.}
\item{source_ids}{Comma-separated string. The Unique Universal Identifier (UUID) of the sources used to build the variable.}
}
}
}
\usage{
pgvariables
}
\description{
These are the variables available in PRIO-GRID.
}
\details{
The static information is used when building PRIO-GRID, as there are different naming conventions
for the data depending on whether the data is static or with a temporal-dimension.
}
\keyword{datasets}
