% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgvariables}
\alias{pgvariables}
\title{PRIO-GRID Variables}
\format{
\subsection{\code{pgvariables}}{

A data frame with 4 rows and 5 columns:
\describe{
\item{name}{Full name of the variable.}
\item{static}{Boolean. Whether the variable varies over time or not.}
\item{source_name}{Full name of the source, preferably including the institution hosting it.}
\item{source_version}{The version of the data source, as noted by the creator. If none, use the publication year.}
\item{source_id}{The Unique Universal Identifier (UUID) of the source}
}
}
}
\usage{
pgvariables
}
\description{
These are the variables available in PRIO-GRID.
}
\details{
The static information is used when building PRIO-GRID, as there are different naming conventions
for the data depending on whether the data is static or with a temporal-dimension.
}
\keyword{datasets}
