% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfunctions.R
\name{prio_blank_grid}
\alias{prio_blank_grid}
\title{Blank GRID
Returns a blank grid with default PRIOGrid dimensions}
\usage{
prio_blank_grid(ncol = FALSE, nrow = FALSE, crs = FALSE, extent = FALSE)
}
\arguments{
\item{ncol}{Number of columns, defaults to prio_ncol}

\item{nrow}{Number of rows, defaults to prio_nrow}

\item{crs}{CRS string, defaults to prio_crs}
}
\value{
A blank grid
}
\description{
Blank GRID
Returns a blank grid with default PRIOGrid dimensions
}
