% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prio_blank_grid}
\alias{prio_blank_grid}
\title{Create a raster with PRIO-GRID ids.}
\usage{
prio_blank_grid(
  ncol = pgoptions$get_ncol(),
  nrow = pgoptions$get_nrow(),
  extent = pgoptions$get_extent(),
  crs_string = pgoptions$get_crs()
)
}
\arguments{
\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{extent}{The extent of the raster as a vector c(xmin, xmax, ymin, ymax)}

\item{crs_string}{A CRS-string. Description of the Coordinate Reference System (map projection) in PROJ.4, WKT or authority:code notation.}
}
\value{
SpatRaster
}
\description{
Create a raster with PRIO-GRID ids.
}
\examples{
pg <- prio_blank_grid()
}
