% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priogrid-package.R
\docType{package}
\name{priogrid-package}
\alias{priogrid}
\alias{priogrid-package}
\title{priogrid: PRIO-GRID Replication Code}
\description{
Codebase for building PRIO-GRID from a diverse set of raw data.
}
\author{
\strong{Maintainer}: Jonas Vestby \email{jonves@prio.org}

Authors:
\itemize{
  \item Andreas Forø Tollefsen \email{andreas@prio.org}
  \item Peder Landsverk \email{pglandsverk@gmail.com}
  \item Vilde villar@prio.org Bergstad Larsen
}

}
\keyword{internal}
