% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rast_to_df}
\alias{rast_to_df}
\title{Converts raster with variable to data.frame}
\usage{
rast_to_df(rast, static = TRUE, varname = NULL)
}
\arguments{
\item{rast}{SpatRaster}

\item{static}{True if no temporal dimension, False else.}

\item{varname}{The variable name, only required if static is False.}
}
\description{
Assumes that the name of the raster layer is the name of the variable if static is true,
otherwise, the user must supply the correct variable name. If static is false,
the name of the raster layer is assumed to be the time variable.
}
\examples{
ne <- gen_naturalearth_cover()
rast_to_df(ne)
}
