% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cshapes.R
\name{read_cshapes}
\alias{read_cshapes}
\title{Reads the CShapes 2.0 raw data}
\usage{
read_cshapes()
}
\value{
an object of class sf
}
\description{
It correctly formats gwsdate and gwedate as date objects, and adds a utility column date_interval.
}
