% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geoEPR.R
\name{read_epr}
\alias{read_epr}
\title{Reads the EPR Core 2023 data}
\usage{
read_epr()
}
\value{
an object of class tibble
}
\description{
Reads the EPR Core 2023 data
}
\references{
\insertRef{vogtIntegratingDataEthnicity2015}{priogrid}
}
