% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geoEPR.R
\name{read_geoepr}
\alias{read_geoepr}
\title{Reads the GeoEPR - Geo-referencing Ethnic Power Relations data}
\usage{
read_geoepr()
}
\value{
an object of class sf
}
\description{
Formats gwsdate and gwedate as date objects, and adds a utility column date_interval
}
\references{
\insertRef{wucherpfennigPoliticallyRelevantEthnic2011}{vogtIntegratingDataEthnicity2015}{priogrid}
}
