% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geopko.R
\name{read_geopko}
\alias{read_geopko}
\title{Reads the Geocoded Peacekeeping Operations (Geo-PKO) data}
\usage{
read_geopko()
}
\value{
an object of class sf
}
\description{
Reads the Geocoded Peacekeeping Operations (Geo-PKO) data
}
