% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ghsl.R
\name{read_ghsl_population_grid}
\alias{read_ghsl_population_grid}
\title{Reads the GHSL GHS Population Grid data}
\usage{
read_ghsl_population_grid()
}
\value{
an object of class sf
}
\description{
Reads the GHSL GHS Population Grid data
}
\references{
\insertRef{schiavinaGHSPOPR2023AGHS2023}{priogrid}
}
