% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naturalearth.R
\name{read_ne_disputed_areas}
\alias{read_ne_disputed_areas}
\title{Read Natural Earth Disputed Areas dataset}
\usage{
read_ne_disputed_areas()
}
\value{
sf data.frame
}
\description{
Reads the Admin 0 – Breakaway, Disputed Areas dataset from Natural Earth.
This dataset shows de facto boundaries based on control versus de jure.
The data entries are not dated.
}
\examples{
# Load disputed areas dataset
disputed_areas <- read_disputed_areas()
plot(sf::st_geometry(disputed_areas), main = "Disputed Areas")
}
\references{
\insertRef{naturalearthAdmin0Breakaway2024}{priogrid}
}
