% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{read_pg_static}
\alias{read_pg_static}
\title{Collects and returns static (non-timevarying) PRIO-GRID data}
\usage{
read_pg_static(as_raster = FALSE, test = FALSE, overwrite = FALSE)
}
\arguments{
\item{as_raster}{Boolean (default FALSE). Set to TRUE to return a list of rasters. Defaults to returning a data.table.}

\item{test}{Boolean. Prints coverage summary (percentage cells with data) for each variable.}

\item{overwrite}{Boolean (default FALSE). Checks if the data.table exists as a file if as_raster = FALSE.
If it does, then it returns the data.table. If not, it saves the data.table to disk as .csv.gz and .parquet.}
}
\value{
a data.table with pgid as rows and variables as columns or a list of terra SpatRasters
}
\description{
See \link{pgvariables} for an overview of variables.
}
\examples{
\dontrun{
  pg_dt <- read_pg_static(test = TRUE) # prints coverage summary for each variable
  pg_rast <- read_pg_static(as_raster = TRUE)
}
}
