% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{read_pg_timevarying}
\alias{read_pg_timevarying}
\title{Collects and returns timevarying (non-static) PRIO-GRID data}
\usage{
read_pg_timevarying(as_raster = FALSE, test = FALSE, overwrite = FALSE)
}
\arguments{
\item{as_raster}{Boolean. Set to TRUE to return a list of rasters. Defaults to FALSE (returns data.table).}

\item{test}{Boolean. Returns only coverage summary (percentage cells with data) for each variable for each date.}

\item{overwrite}{Boolean (default FALSE). Checks if the data.table exists as a file if as_raster = FALSE.
If it does, then it returns the data.table. If not, it saves the data.table to disk as .csv.gz and .parquet.}
}
\value{
a data.table with pgid + measurement_date as rows and variables as columns or a list of terra SpatRasters with each measurement date as layers or coverage test
}
\description{
See \link{pgvariables} for an overview of variables.
}
\examples{
\dontrun{
  pg_dt <- read_pg_timevarying() # data.table
  coverage_test <- read_pg_timevarying(test = TRUE) # coverage test data frame
  pg_rast <- read_pg_timevarying(as_raster = TRUE) # list of rasters
}
}
