% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shdi_data.R
\name{read_shdi}
\alias{read_shdi}
\title{Read the GDL SHDI data}
\usage{
read_shdi()
}
\value{
A tibble with SHDI data.
}
\description{
Reads the downloaded SHDI dataset from GlobalDataLab.
}
\references{
\insertRef{globaldatalabSubnationalHumanDevelopment2019}{priogrid}
}
