% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_speibase.R
\name{read_speibase}
\alias{read_speibase}
\title{Read and Process Global SPEI database data}
\usage{
read_speibase(interval = 6)
}
\arguments{
\item{interval}{Integer. The month interval to calculate SPEI over. E.g., if the interval is
6, then monthly precipitation and potential evapotranspiration are aggregated over the last 6 months,
then the SPEI (anomaly) is calculated. This is done for each month.}
}
\description{
While this data can be downloaded from the Global SPEI database, it is only open
for download through manual operations on the web-page. The code for reproducing
SPEIbase is open, however. We use this here to process CRU PET and PRE data. The
input for SPEIbase 2.11 is CRU v.4.09.
}
\details{
We use R-code from Beguería S. (2017) SPEIbase: R code used in generating the SPEI global database, doi:10.5281/zenodo.834462.
}
\references{
\insertRef{begueriaSPEIbase2024}{priogrid}
\insertRef{begueriaStandardizedPrecipitationEvapotranspiration2014}{priogrid}
}
\seealso{
\code{\link{get_pgfile}} for file retrieval functionality,
\code{\link{pg_date_intervals}} for PRIO-GRID temporal boundaries
}
