% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_traveltime.R
\name{read_traveltime}
\alias{read_traveltime}
\title{Reads the Estimated Travel Time to Major Cities version year 2000 data
Unzips the folder and imports the .tif file as SpatRaster}
\usage{
read_traveltime()
}
\value{
an object of class SpatRast
}
\description{
Reads the Estimated Travel Time to Major Cities version year 2000 data
Unzips the folder and imports the .tif file as SpatRaster
}
\references{
\insertRef{nelsonTravelTimeMajor2008}{priogrid}
}
