% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{refToRaster}
\alias{refToRaster}
\title{Old grid to raster}
\usage{
refToRaster(timeseries, year = 1970, variable = "gwno")
}
\arguments{
\item{timeseries}{The data frame contained in the timeseries version of PrioGrid}

\item{year}{The desired year. Currently only yields a single raster for a single year.}

\item{variable}{A vector of variable[s] to return in a stack}
}
\value{
A raster stack with rasters for each variable
}
\description{
Converts the .csv PrioGrid v.1 into a raster.
Useful for testing equivalency.
}
\examples{
prioTimeseries <- read.csv('timeseries.csv') 
prioStack <- refToRaster(prioTimeseries, year = 1993, variable = 'gwno')
plot(prioStack)
}
