% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ruggedterrain.R
\name{ruggedterrain_variable}
\alias{ruggedterrain_variable}
\title{Calculate elevation in PRIO-GRID format with set variable}
\usage{
ruggedterrain_variable(variable)
}
\arguments{
\item{variable}{Variable indicating elevation function.
"elevation_min", "elevation_max", "elevation_mean", "elevation_sdev"}
}
\value{
SpatRast
}
\description{
Takes either the min, max, mean or sdev of elevation in each PRIO-GRID cell
}
\details{
Returns the value of elevation
}
\examples{
# rt <- ruggedterrain_variable(variable = "elevation_mean")


}
\references{
\insertRef{danielsonGlobalMultiresolutionTerrain2011}{priogrid}
}
