% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ruggedterrain.R
\name{ruggedterrain_variable}
\alias{ruggedterrain_variable}
\title{Calculate elevation in PRIO-GRID format with set variable}
\usage{
ruggedterrain_variable(variable)
}
\arguments{
\item{variable}{Character string indicating elevation function.
Must be one of: "elevation_min", "elevation_max", "elevation_mean"}
}
\value{
A \code{SpatRaster} object
}
\description{
Estimates either the min, max, or mean of elevation in each PRIO-GRID cell
based on vector data from GMTED2010.
}
\examples{
\dontrun{
# Get mean elevation for each PRIO-GRID cell
mean_elev <- ruggedterrain_variable(variable = "elevation_mean")

# Get minimum elevation for each PRIO-GRID cell
min_elev <- ruggedterrain_variable(variable = "elevation_min")

# Get maximum elevation for each PRIO-GRID cell
max_elev <- ruggedterrain_variable(variable = "elevation_max")
}

}
\references{
\insertRef{danielsonGlobalMultiresolutionTerrain2011}{priogrid}
}
