% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_priogrid.R
\name{save_pgvariable}
\alias{save_pgvariable}
\title{Save a PRIO-GRID variable}
\usage{
save_pgvariable(rast, varname)
}
\arguments{
\item{rast}{Terra SpatRast from a gen_ function}

\item{varname}{Character string with the variable name}
}
\description{
Save a PRIO-GRID variable to a standard folder depending on your pgoptions:
rawfolder/priogrid/(package version)/(spatial options hash)/(temporal options hash)
}
\details{
Correctly wrap the terra object before writing it, and checks if the variable
is included in the list of variables in PRIOGRID \link{pgvariables}.
}
